pentesting_information: split the 3298-line class into topic mixins (A.2)
Behavior-preserving structural refactor. PenTestingInformation was one 3298-line class with
74 methods; it is now composed from focused mixins in the same package:
- _pentesting_core.py (PenTestingCore): __init__, PromptPurpose->generate_* dispatch
  (explore_steps), setup, and the shared endpoint/schema/account helpers
- _pentesting_auth.py, _pentesting_authz.py, _pentesting_injection.py, _pentesting_session.py,
  _pentesting_ratelimit.py, _pentesting_misc.py: each holds one topic's generate_*_prompts and
  its test_* scenario builders

pentesting_information.py is now a 28-line module composing the mixins and keeping the public
PenTestingInformation name/import path. Mixins share `self`, so cross-topic helper calls resolve
unchanged. The split was done programmatically with asserts that every method is assigned exactly
once; verified all 74 original methods are present on the composed class and the suite is
unchanged (51 passed, 1 skipped). Largest file is now ~816 lines (was 3298).
