+++ src/hackingBuddyGPT/utils/prompt_generation/information/_pentesting_auth.py
@@ -1,0 +1,584 @@
+# This file is part of the PenTestingInformation split (see pentesting_information.py).
+import base64
+import copy
+import csv
+import glob
+import os
+import random
+import re
+import secrets
+from typing import List
+
+from hackingBuddyGPT.utils.openapi.openapi_parser import OpenAPISpecificationParser
+from hackingBuddyGPT.utils.prompt_generation.information.prompt_information import (
+    PromptPurpose,
+)
+from faker import Faker
+
+
+class AuthScenarios:
+    """AuthScenarios: scenario/helper methods mixed into PenTestingInformation."""
+
+    def generate_authentication_prompts(self):
+        """
+            Generate a list of prompts for testing authentication mechanisms on protected endpoints.
+
+            This function constructs test prompts for various authentication scenarios, including:
+            - Accessing protected endpoints with different user accounts.
+            - Using login credentials to acquire tokens.
+            - Testing endpoints that require path parameters like user IDs.
+            - Verifying refresh token mechanisms if applicable.
+
+            Returns:
+                list: A list of prompts for testing authentication and authorization.
+            """
+        prompts = []
+        endpoints = self.get_correct_endpoints_for_method("protected_endpoint", "GET")
+
+        prompts = self.resource_endpoints(prompts)
+
+        if len(endpoints) != 0:
+            for endpoint, login in zip(endpoints, self.login_endpoint):
+                for account in self.accounts:
+                    endpoint_dict = endpoint
+                    if isinstance(endpoint, dict):
+                        endpoint = endpoint.get("path")
+
+                    login_path, login_schema = self.get_path_and_schema(login)
+                    if login_schema is None:
+                        continue
+                    login_schema = self.fill_in_schema_with_account(login_schema, account)
+
+                    if "api" in endpoint and len(endpoint.split("/")) > 0 and "api" in account:
+                        if account["api"] in endpoint:
+                            prompts = self.test_authentication(endpoint, account, prompts)
+                    else:
+                        prompts = self.test_authentication(endpoint, account, prompts)
+                    if "_id}" in endpoint:
+                        endpoint = self.replace_id_placeholder(endpoint, "1")
+
+                    if login_path:
+
+                        if "api" in endpoint and len(endpoint.split("/")) > 0:
+                            if account["api"] in endpoint:
+                                id = account.get("id")
+                                if id and "{id}" in endpoint:
+                                    new_endpoint = endpoint.replace("{id}", str(account.get("id")))
+                                    prompts = self.test_token(login_path, new_endpoint, account, login_schema, prompts)
+                                    prompts = self.random_common_users(new_endpoint, login_path, login_schema, prompts)
+                                else:
+
+                                    prompts = self.test_token(login_path, endpoint, account, login_schema, prompts)
+                                    prompts = self.random_common_users(endpoint, login_path, login_schema, prompts)
+
+                        else:
+                            if "id}" in endpoint:
+                                endpoint = self.replace_placeholders_with_1(endpoint, f"{account.get('id')}")
+                                prompts = self.random_common_users(endpoint, login_path, login_schema, prompts)
+                                prompts = self.test_token(login_path, endpoint, account, login_schema, prompts)
+
+            if self.current_refresh_endpoint:
+                refresh_get_endpoints = self.get_correct_endpoints_for_method("refresh_endpoint", "GET")
+                refresh_post_endpoints = self.get_correct_endpoints_for_method("refresh_endpoint", "POST")
+                if len(refresh_get_endpoints) != 0 and refresh_post_endpoints:
+                    for account in self.accounts:
+
+                        for refresh_get_endpoint, refresh_post_endpoint in zip(refresh_get_endpoints,
+                                                                               refresh_post_endpoints):
+                            if "id}" in refresh_get_endpoint:
+                                refresh_get_endpoint = self.replace_placeholders_with_1(refresh_get_endpoint,
+                                                                                        account.get("id"))
+                                if account["api"] in refresh_get_endpoint:
+                                    prompts = self.test_refresh_token(refresh_post_endpoint, refresh_get_endpoint,
+                                                                      account, prompts)
+                                else:
+                                    prompts = self.test_refresh_token(refresh_post_endpoint, refresh_get_endpoint,
+                                                                      account, prompts)
+
+        return prompts
+
+    def generate_special_authentication(self):
+        """
+        Generate prompts for advanced authentication testing.
+
+        This method focuses on testing the security robustness of login and authentication mechanisms.
+        It includes:
+        - Injection tests with special accounts (e.g., "bender" accounts).
+        - Brute force protection validation using repeated login attempts.
+        - CSS (Content Sniffing) vulnerability checks on authentication endpoints.
+        - Token revocation checks after login.
+        - ID substitution for user-specific paths.
+        - Custom test cases for comments and advanced login workflows.
+
+        Returns:
+            list: A list of prompts targeting special authentication test cases.
+        """
+        prompts = []
+
+        self.counter = self.counter + 1
+
+        for login in self.login_endpoint:
+
+            login_path, login_schema = self.get_path_and_schema(login)
+
+            prompts = self.test_injection_bender_account(login_path, prompts)
+
+            if login_path:
+                prompts = self.test_brute_force(login_path, login_schema, prompts)
+
+            for acc in self.accounts:
+
+                if "api" in login_path and "api" in acc.keys():
+                    if acc["api"] in login_path:
+                        str_id = f"{acc.get('id')}"
+                        login_path = login_path.replace("{id}", str_id)
+
+                        str = [
+                                  f"Attempt Login with Invalid Credentials - Send repeated POST requests to the login endpoint {login_path} using invalid credentials (e.g., incorrect password) {self.get_credentials(login_schema, login_path)}. " +
+                                  f"Simulate a high rate of attempts (e.g., 10 requests within 1 minute) to trigger potential brute force protection mechanisms. "] * 10
+
+                        if login_path:
+                            prompts.append(  # Brute Force Protection
+                                {
+                                    "objective": "Brute Force Protection",
+                                    "steps": str,
+                                    "expected_response_code":
+                                        [[
+                                            "429 Too Many Requests to indicate effective rate-limiting measures against brute force attacks.",
+                                            "403 Forbidden as an alternative response indicating account lockout or other security measures preventing further attempts."
+                                        ]] * 10,
+                                    "token": [""] * 10,
+                                    "path": [login_path] * 10,
+                                    "security":
+                                        ["Ensure that user cannot login with invalid credentials.\n" +
+                                         "Ensure that rate-limiting mechanisms are robust and properly configured to prevent brute force attacks by limiting the number of allowed failed attempts within a given time frame. This prevents attackers from trying a large number of combinations in a short period.\n" +
+                                         "Check that account lockout mechanisms or other access denial policies are effective in disabling further login attempts after a certain number of failures, protecting against continuous brute force attempts and securing user accounts from unauthorized access."]
+
+                                }
+                            )
+                        if self.auth_endpoint:
+                            if acc["api"] in login_path:
+                                str_id = f"{acc.get('id')}"
+                                login_path = login_path.replace("{id}", str_id)
+
+                            get_paths = self.get_correct_endpoints_for_method("auth_endpoint", "GET")
+                            post_paths = self.get_correct_endpoints_for_method("auth_endpoint", "POST")
+
+                            for get_path in get_paths:
+                                if acc["api"] in get_path:
+                                    str_id = f"{acc.get('id')}"
+                                    get_path = get_path.replace("{id}", str_id)
+                                    prompts = self.test_css(get_path, prompts)
+
+                            for post_path in post_paths:
+                                if acc["api"] in post_path:
+                                    str_id = f"{acc.get('id')}"
+                                    post_path = post_path.replace("{id}", str_id)
+                                    schema = self.openapi_spec_parser.get_schema_for_endpoint(post_path, "POST")
+                                    prompts = self.test_css(post_path, prompts, schema=schema)
+
+                            if self.current_protected_endpoint:
+                                get_endpoints = self.get_correct_endpoints_for_method("protected_endpoint", "GET")
+
+                                for get_endpoint in get_endpoints:
+                                    for account in self.accounts:
+                                        if acc["api"] in get_endpoint:
+                                            str_id = f"{acc.get('id')}"
+                                            get_endpoint = get_endpoint.replace("{id}", str_id)
+                                        prompts = self.test_token(login_path, get_endpoint, account, login_schema,
+                                                                  prompts, revocation=True)
+
+
+                else:
+
+                    if login_path:
+                        prompts = self.test_brute_force(login_path, login_schema, prompts, number=10)
+
+                        if self.auth_endpoint:
+
+                            get_paths = self.get_correct_endpoints_for_method("auth_endpoint", "GET")
+                            post_paths = self.get_correct_endpoints_for_method("auth_endpoint", "POST")
+
+                            for get_path in get_paths:
+                                prompts = self.test_css(get_path, prompts)
+
+                            for post_path in post_paths:
+                                schema = self.openapi_spec_parser.get_schema_for_endpoint(post_path, "POST")
+                                prompts = self.test_css(post_path, prompts, schema=schema)
+
+                        if self.current_protected_endpoint:
+                            get_endpoints = self.get_correct_endpoints_for_method("protected_endpoint", "GET")
+
+                            for get_endpoint in get_endpoints:
+                                for account in self.accounts:
+                                    if "id}" in get_endpoint:
+                                        get_endpoint = self.replace_placeholders_with_1(get_endpoint, account.get("id"))
+
+                                    prompts = self.test_token(login_path, get_endpoint, account, login_schema, prompts,
+                                                              revocation=True)
+
+                    # return prompts
+
+                prompts = self.test_comment(acc, prompts)
+
+
+                return prompts
+
+        return prompts
+
+    def test_authentication(self, endpoint, account, prompts):
+        if "api" in endpoint:
+
+            new_endpoint, endpoint_of_other_user = self.replace_ids(account, endpoint)
+
+        else:  # Non-API endpoints
+            new_endpoint, endpoint_of_other_user = self.replace_ids(account, endpoint)
+        prompts = prompts + [  # Check Access Without Authentication
+            {
+                "objective": "Check Access Without Authentication",
+                "steps": [
+                    f"make a GET request to the protected endpoint {new_endpoint} without including any authentication headers with user:{account}.\n"],
+                "expected_response_code": ["401 Unauthorized", "403 Forbidden."],
+                "token": [""],
+                "path": [new_endpoint],
+                "security": [
+                    "Ensure that the correct expected code is received  confirming that access is restricted and no sensitive data is included in the response body."]},
+            # Test Valid Authentication
+            {
+                "objective": "Test Valid Authentication",
+                "steps": [
+                    f"use this accounts token to send a GET request to {new_endpoint}."],
+                "expected_response_code": [
+                    "200 OK response, indicating successful authentication."
+                ],
+                "path": [new_endpoint],
+                "token": [account.get("token")],
+                "security": ["Verify that the response includes the expected data.",
+                             "Confirm that the token provides the correct access level."]
+
+            },
+            # Test only can access my user
+            {
+                "objective": "Test cannot access other user data",
+                "steps": [
+                    f"use this accounts token to send a GET request to {endpoint_of_other_user}."],
+                "expected_response_code": ["401 Unauthorized", "403 Forbidden."],
+                "token": [account.get("token")],
+                "path": [new_endpoint],
+                "security": [
+                    "Ensure that the correct expected code is received  confirming that access is restricted and no sensitive data is included in the response body."]},
+        ]
+
+        return prompts
+
+    def test_token(self, login_path, endpoint, account, login_schema, prompts, revocation=False):
+        if revocation and endpoint is not None:
+
+            prompts.append(
+
+                # Token Revocation
+                {
+                    "objective": "Token Revocation",
+                    "steps": [
+                        f"Send a GET request to {endpoint} with a valid token = {account.get('token')}.",
+                        f"Send a POST request to the token revocation endpoint {login_path} to invalidate an active token with user credentials:{account}",
+                        f"Test the revoked token by sending a GET request to {endpoint} with the revoked token."
+                    ],
+                    "expected_response_code": [
+                        "200 OK for the initial use of the valid token, confirming it is active.",
+                        "200 OK or 204 No Content, Successful revocation "
+                        "401 Unauthorized for attempts made with the revoked token, confirming the system's capability to recognize and deny access with revoked credentials."
+                    ],
+                    "token": [account.get("token"), "", account.get("token")],
+                    "path": [endpoint, login_path, endpoint],
+                    "security":
+                        [
+                            "Ensure that valid tokens are correctly recognized and granted access, confirming that authentication systems are functioning as expected before revocation.",
+                            "Verify that once tokens are revoked, the system immediately denies access, preventing any security breaches that could occur from the continued use of revoked tokens. This includes ensuring timely updates across all components of the system that might interact with token validation processes."]
+
+                }
+
+            )  # protected end point needed
+        else:
+            prompts = prompts + [
+                {  # Test Expired or Invalid Tokens
+                    "objective": "Test Expired or Invalid Tokens",
+                    "steps": [
+                        f"Send a POST request to the login endpoint {login_path} with valid login credentials ({login_schema}) to obtain a valid token with user:{account}.\n",
+                        f"Send a GET request to {endpoint} using an expired, revoked, or otherwise invalid token."
+                    ],
+                    "path": [login_path, endpoint],
+                    "expected_response_code": [
+                        "200 OK for successfully obtaining a valid token with correct credentials.",
+                        "401 Unauthorized for the expired or invalid token, verifying that the system rejects tokens that are no longer valid."
+                    ],
+                    "token": ["", self.create_random_bearer_token()],
+                    "security": [
+                        "Ensure that the system appropriately validates token expiration or invalidation before granting access.",
+                        "Ensure that no sensitive information is included in the error responses."
+                    ]
+                },
+                # Basic Authentication (if applicable)
+                {
+                    "objective": "Valid Basic Authentication",
+                    "steps": [
+                        f"Send a GET request to {endpoint} using Basic Authentication with the valid token {self.valid_token}."
+                    ],
+                    "path": [endpoint],
+                    "expected_response_code": [
+                        "200 OK when sending a GET request with correct credentials, confirming access is granted."
+                    ],
+                    "token": [account.get("token")],
+                    "security": [
+                        "Ensure the system returns a 200 OK response for valid credentials and does not expose sensitive data in the response body."
+                    ]
+                },
+                {
+                    "objective": "Invalid Basic Authentication",
+                    "steps": [
+                        f"Attempt to send a POST request to the login endpoint {login_path} using invalid login credentials {login_schema} with user:{account}.\n",
+                        f"Send a GET request to {endpoint} using Basic Authentication with invalid credentials:{self.get_invalid_credentials(account)}"
+                    ],
+                    "path": [login_path, endpoint],
+                    "expected_response_code": [
+                        "401 Unauthorized when attempting to authenticate with invalid credentials.",
+                        "401 Unauthorized when attempting to access the protected endpoint with invalid credentials."
+                    ],
+                    "token": [account.get("token"), account.get("token")],
+                    "security": [
+                        "Ensure that the system rejects invalid credentials with a 401 Unauthorized status.",
+                        "Ensure no sensitive information is leaked in the error responses, such as specific reasons for the failure."
+                    ]
+                }
+                ,
+
+            ]
+
+        return prompts
+
+    def test_refresh_token(self, refresh_post_endpoint, refresh_get_endpoint, account, prompts):
+        prompts = prompts + [  # Test Token Refresh (if applicable)
+
+            {
+                "objective": "Test Token Refresh",
+                "steps": [
+                    f"send a GET request to {refresh_get_endpoint} with the expired token in the Authorization header. Verify that the API responds with a 401 Unauthorized status, indicating the token has expired.",
+                    f"send a POST request to the token refresh endpoint {refresh_post_endpoint} with the valid refresh token in the request body or headers, depending on the API's token refresh requirements. Check if the API responds with a 200 OK status and includes a new access token in the response body.",
+                    f"use the new access token to send a GET request to {refresh_get_endpoint} again. Confirm that the API responds with a 200 OK status, indicating successful access with the refreshed token, and that the old expired token is no longer valid."
+                ],
+                "path": [refresh_get_endpoint, refresh_get_endpoint, refresh_get_endpoint],
+                "token": [self.create_random_bearer_token(),
+                          account.get("token"),
+                          account.get("token")],
+                "expected_response_code": [
+                    "401 Unauthorized for the expired token use, verifying that the token has indeed expired and is recognized by the system as such.",
+                    "200 OK upon refreshing the token, confirming that the refresh mechanism works as expected and a new token is issued correctly.",
+                    "200 OK when using the new token, verifying that the new token grants access and the old token is invalidated."
+                ],
+                "security": [
+                    "Ensure that the API does not leak sensitive information in error responses and that expired tokens are promptly invalidated to prevent unauthorized use."]
+            }
+
+            # This prompt tests if the API correctly handles token expiration and issues a new token upon refresh,
+            # while ensuring that the expired token no longer provides access to protected resources.
+
+        ]
+        return prompts
+
+    def mechanic_report(self, endpoint, account, prompts):
+        if "mechanic_report" in endpoint:
+            users = list(range(1, 100))
+            random_numbers = random.sample(users, 10)
+            random_numbers = [5] + random_numbers
+
+            for i in random_numbers:
+                new_endpoint = endpoint + f"?report_id={i}"
+                prompts.append(
+
+                    # Access Control to Specific Resources
+                    {
+                        "objective": "Access Control to Specific Resource of other used",
+                        "steps": [
+                            f"Attempting Access with Non-Owner's Credentials - Send a GET request to {new_endpoint} using an authentication token for User B (a different user who does not own the resource). user:{account}.\n",
+                        ],
+                        "path": [new_endpoint],
+                        "expected_response_code": [
+                            "403 Forbidden or 404 Not Found when accessed by a non-owner, confirming that access is appropriately restricted.",
+
+                        ],
+                        "token": [account.get("token")],
+                        "security": [
+                            "Ensure robust validation of user credentials and proper enforcement of access controls. Verify that the system correctly identifies ownership and prevents unauthorized access based on user authentication and authorization levels."
+                            ]
+                    }
+
+                    # This prompt tests access control for resources that are user-owned by verifying:
+                    # - Only the resource owner can access the data (User A)
+                    # - Non-owners receive a restricted response (User B)
+                    # - Unauthenticated requests are denied access (if applicable)
+                )
+        return prompts
+
+    def random_common_users(self, endpoint, login_path, login_schema, prompts):
+        random_entries = random.sample(self._credentials)
+        if len(random_entries) >= 10:
+            random_entries = random_entries[:10]
+
+        for index, random_entry in random_entries:
+            username = random_entry['username']
+            password = random_entry['password']
+            # Now you can print or use username and password as needed
+            common_account = {"email": username, "password": password}
+            prompts = prompts + [{
+                "objective": "Attempt login with common username and password",
+                "steps": [
+                    f"Attempt to send a POST request to the login endpoint {login_path} using common login credentials {login_schema} with user:{common_account}.\n", ],
+                "path": [login_path],
+                "expected_response_code": [
+                    "401 Unauthorized when attempting to authenticate with invalid credentials.",
+                ],
+                "token": [""],
+                "security": [
+                    "Ensure that the system rejects invalid credentials with a 401 Unauthorized status.",
+                    "Ensure no sensitive information is leaked in the error responses, such as specific reasons for the failure."
+                ]
+            }]
+        username = "savanna48@ortiz.com"
+        password = "zTyBwV/9"
+        common_account = {"email": username, "password": password}
+        prompts = prompts + [{
+            "objective": "Attempt login with common username and password",
+            "steps": [
+                f"Attempt to send a POST request to the login endpoint {login_path} using common login credentials {login_schema} with user:{common_account}.\n",
+            ],
+            "path": [login_path],
+            "expected_response_code": [
+                "401 Unauthorized when attempting to authenticate with invalid credentials.",
+            ],
+            "token": [""],
+            "security": [
+                "Ensure that the system rejects invalid credentials with a 401 Unauthorized status.",
+                "Ensure no sensitive information is leaked in the error responses, such as specific reasons for the failure."
+            ]
+        }]
+
+        return prompts
+
+    def test_brute_force(self, login_path, login_schema, prompts, number=100):
+        if number == 10:
+
+            str = [
+                      f"Attempt Login with Invalid Credentials - Send repeated POST requests to the login endpoint {login_path} using invalid credentials (e.g., incorrect password) {self.get_credentials(login_schema, login_path)}. " +
+                      f"Simulate a high rate of attempts (e.g., 10 requests within 1 minute) to trigger potential brute force protection mechanisms. "] * 10
+
+            prompts.append(  # Brute Force Protection
+                {
+                    "objective": "Brute Force Protection",
+                    "steps": str,
+                    "expected_response_code":
+                        [[
+                            "429 Too Many Requests to indicate effective rate-limiting measures against brute force attacks.",
+                            "403 Forbidden as an alternative response indicating account lockout or other security measures preventing further attempts."
+                        ]] * 10,
+                    "token": [""] * 10,
+                    "path": [login_path] * 10,
+                    "security":
+                        ["Ensure that user cannot login with invalid credentials.\n" +
+                         "Ensure that rate-limiting mechanisms are robust and properly configured to prevent brute force attacks by limiting the number of allowed failed attempts within a given time frame. This prevents attackers from trying a large number of combinations in a short period.\n" +
+                         "Check that account lockout mechanisms or other access denial policies are effective in disabling further login attempts after a certain number of failures, protecting against continuous brute force attempts and securing user accounts from unauthorized access."]
+
+                }
+            )
+        else:
+            if len(self.brute_force_accounts) > 0:
+                for acc_b in self.brute_force_accounts:
+                    acc_b["x"] = 100
+                    str = [
+                              f"Attempt Login with different Credentials - Send repeated POST requests to the login endpoint {login_path} using invalid credentials (e.g., incorrect password) user:{acc_b}.\n" +
+                              f"Simulate a high rate of attempts (e.g., 10 requests within 1 minute) to trigger potential brute force protection mechanisms. "] * 10
+
+                    prompts.append(  # Brute Force Protection
+                        {
+                            "objective": "Brute Force Attack",
+                            "steps": str,
+                            "expected_response_code":
+                                [
+                                    "429 Too Many Requests to indicate effective rate-limiting measures against brute force attacks.",
+                                    "403 Forbidden as an alternative response indicating account lockout or other security measures preventing further attempts."
+                                ] * 10,
+                            "token": [""] * 10,
+                            "path": [login_path] * 10,
+                            "security":
+                                ["Ensure that user cannot login with invalid credentials.\n" +
+                                 "Ensure that rate-limiting mechanisms are robust and properly configured to prevent brute force attacks by limiting the number of allowed failed attempts within a given time frame. This prevents attackers from trying a large number of combinations in a short period.\n" +
+                                 "Check that account lockout mechanisms or other access denial policies are effective in disabling further login attempts after a certain number of failures, protecting against continuous brute force attempts and securing user accounts from unauthorized access."]
+
+                        }
+                    )
+
+            change_email = self.extract_change_email_endpoints_with_schema()
+            for endpoint, schema in change_email:
+                for account in self.accounts:
+                    if isinstance(endpoint, dict):
+                        schema = endpoint.get("schema")
+                        schema = self.fill_schema(schema)
+                    else:
+                        schema = self.fill_schema(schema)
+                        prompts.append(  # Insecure Email Change Process
+                            {
+                                "objective": "Insecure Email Change Process",
+                                "steps": [f"Send a POST request to endpoint {endpoint} with schema {schema}"],
+                                "expected_response_code":
+                                    [
+                                        "403 Forbidden if proper authorization checks prevent changing another user's email."
+                                    ],
+                                "token": [account.get("token")],
+                                "path": [endpoint],
+                                "security":
+                                    [
+                                        "Ensure that users cannot change the email address of other users without proper authorization.\n" +
+                                        "Verify that sensitive actions like email changes send verification tokens only to the current, verified email address of the account owner.\n" +
+                                        "Check for IDOR vulnerabilities and enforce proper ownership validation when modifying sensitive user information."]
+                            }
+                        )
+
+            otp_endpoints = [ep for ep in self.endpoints if "otp"]
+            reset_password = [ep for ep in self.endpoints if
+                              "reset" or "password" or "forget-password" or "forgot" in ep]
+            otps = list(range(0000, 9999))
+            random_otps = [8859] + random.sample(otps, 10)
+            otp_calls = ""
+            counter = 0
+            otps = []
+            for (otp, otp_endpoint) in zip(random_otps, otp_endpoints):
+                otp_request = {"email": "adam007@example.com", "otp": f'{otp}', "password": "Az123###"}
+
+                otp_calls = otp_calls + f"Send a POST request to endpoint {otp_endpoint} with the following body {otp_request}.\n"
+                counter += 1
+                otps.append(otp)
+
+            for reset in reset_password:
+                user = {"email": "adam007@example.com"}
+                # triggers otp being sent
+                new_otp_calls = f"Send a POST request to endpoint{reset} for victim user:{user}.\n" + otp_calls
+                prompts.append(  # OTP Bruteforce Protection
+                    {
+                        "objective": "OTP Bruteforce Protection",
+                        "steps": [new_otp_calls],
+                        "expected_response_code":
+                            [
+                                "429 Too Many Requests indicating effective rate-limiting against OTP bruteforce attempts.",
+                                "403 Forbidden indicating account lockout or blocking after multiple failed OTP attempts."
+                            ] * counter,
+                        "token": [""] * counter,
+                        "path": [reset] + otps,
+                        "security":
+                            [
+                                "Ensure that OTP verification mechanisms enforce rate-limiting to prevent brute-force attacks.\n" +
+                                "Verify that deprecated or older API versions do not bypass security controls.\n" +
+                                "Ensure consistent security protections across all API versions to prevent attackers from exploiting outdated endpoints."]
+                    }
+                )
+
+        return prompts
+
