+++ src/hackingBuddyGPT/utils/prompt_generation/information/_pentesting_authz.py
@@ -1,0 +1,495 @@
+# This file is part of the PenTestingInformation split (see pentesting_information.py).
+import base64
+import copy
+import csv
+import glob
+import os
+import random
+import re
+import secrets
+from typing import List
+
+from hackingBuddyGPT.utils.openapi.openapi_parser import OpenAPISpecificationParser
+from hackingBuddyGPT.utils.prompt_generation.information.prompt_information import (
+    PromptPurpose,
+)
+from faker import Faker
+
+
+class AuthorizationScenarios:
+    """AuthorizationScenarios: scenario/helper methods mixed into PenTestingInformation."""
+
+    def generate_authorization_prompts(self):
+        """
+        Generate prompts to test authorization and role-based access control (RBAC)
+        on protected endpoints.
+
+        This method performs a series of authorization tests by:
+        - Generating prompts for GET requests to protected endpoints.
+        - Including user-specific endpoints like "users".
+        - Excluding endpoints related to "community".
+        - Testing RBAC behavior across different roles (admin, user, guest).
+        - Verifying proper handling of user-specific data and ID placeholders.
+        - Checking for data masking mechanisms.
+        - Running CRUD operation tests on relevant endpoints.
+
+        Returns:
+            list: A list of authorization prompts for validating access control.
+        """
+        prompts = []
+        endpoints = self.get_correct_endpoints_for_method("protected_endpoint", "GET")
+
+        users_ep = [ep + "s" for ep in self.endpoints if ep.endswith("user")]
+        endpoints = endpoints + users_ep
+        endpoints = [ep for ep in endpoints if "community" not in ep]
+        if len(endpoints) != 0:
+            for endpoint in endpoints:
+                for account in self.accounts:
+                    if isinstance(endpoint, dict):
+                        endpoint = endpoint.get("path")
+                        if "api" in endpoint and len(
+                                endpoint.split("/")) > 0 and "id" in endpoint and not "identity" in endpoint:
+
+                            if "api" in account and account["api"] in endpoint:
+
+                                prompts = self.resource_prompts(endpoint, account, prompts)
+
+                                prompts = self.mechanic_report(endpoint, account, prompts)
+                                if "{id}" in endpoint:
+                                    id = account.get("id")
+                                    prompts = self.rbac(endpoint, account, prompts, id=id)
+                                else:
+
+                                    prompts = self.rbac(endpoint, account, prompts)
+
+                                # Verify Data Masking
+                                prompts = self.verify_data_masking(endpoint, account, prompts)
+
+
+                        else:
+
+                            prompts = self.resource_prompts(endpoint, account, prompts)
+
+                            prompts = self.mechanic_report(endpoint, account, prompts)
+                            if "_id}" in endpoint:
+                                if "id" in account:
+                                    id = account.get("id")
+                                else:
+                                    id = 1
+                                endpoint = self.replace_id_placeholder(endpoint, str(id))
+
+                            if self.admin and self.guest:
+                                prompts.append(
+
+                                    # Verify Role-Based Access Control (RBAC)
+
+                                    {
+                                        "objective": "Verify Role-Based Access Control (RBAC)",
+                                        "steps": [
+                                            f"send a GET request to {endpoint} with an 'admin' role authentication token in the Authorization header. Expect a 200 OK response, indicating access to all resources. Verify that the response data includes complete access to the endpoint, reflecting the admin’s permissions.user:{account}.\n",
+                                            f"send a GET request to {endpoint} with a 'user' role authentication token in the Authorization header. Expect a 200 OK response, but confirm that only authorized resources or data for a regular user are included in the response. Verify that any restricted admin-only resources are not accessible to the user role.user:{account}.\n",
+                                            f"send a GET request to {endpoint} with a 'guest' role authentication token (or without an authentication token if guests are unauthenticated). Expect either a 403 Forbidden or a 404 Not Found response if guests are not permitted to access the endpoint, or a 200 OK with limited data if some resources are publicly accessible to guests.user:{account}.\n"
+                                        ],
+                                        "path": [endpoint, endpoint, endpoint],
+                                        "expected_response_code": [
+                                            "200 OK for admin, confirming full access.",
+                                            "200 OK for users, confirming access is limited to non-admin resources.",
+                                            "403 Forbidden or 404 Not Found for guests if no access is granted, or 200 OK if limited access is applicable."
+                                        ],
+                                        "token": [self.admin.get("token"), account.get("token"),
+                                                  self.guest.get("token")],
+                                        "security": [
+                                            "Verify that the response data includes complete access to the endpoint, reflecting the admin’s permissions.",
+                                            "Verify that any restricted admin-only resources are not accessible to the user role.",
+                                            "Verify that guest role has no or limited access."],
+
+                                    }
+
+                                    # This prompt tests the RBAC implementation by confirming that each user role has the appropriate access level:
+                                    # - Admin has full access
+                                    # - User has restricted access
+                                    # - Guest has no or limited access.
+
+                                )
+
+                            else:
+                                prompts = self.rbac(endpoint, account, prompts)
+
+                            # Verify Data Masking
+                            prompts = self.verify_data_masking(endpoint, account, prompts)
+
+            prompts = self.test_crud(endpoints, prompts)
+        return prompts
+
+    def rbac(self, old_endpoint, account, prompts, id=None):
+        endpoint = copy.deepcopy(old_endpoint)
+
+        ids = [35] + list(range(1, 10))
+        if id:
+            ids = [id] + ids
+
+        endpoint_and_new_endpoint = []
+        for i in ids:
+            token = "" if "login" not in endpoint else account.get("token")
+            endpoint, endpoint_of_other_user = self.replace_ids(account, endpoint, given_id=i)
+            if (endpoint, endpoint_of_other_user) in endpoint_and_new_endpoint:
+                continue
+            endpoint_and_new_endpoint.append((endpoint, endpoint_of_other_user))
+
+            prompts = prompts + [
+
+                # Verify Role-Based Access Control (RBAC)
+
+                {
+                    "objective": "Verify Role-Based Access Control (RBAC)",
+                    "steps": [
+                        # f"send a GET request to {endpoint} with an 'admin' role authentication token in the Authorization header. Expect a 200 OK response, indicating access to all resources. Verify that the response data includes complete access to the endpoint, reflecting the admin’s permissions.",
+                        f"send a GET request to {endpoint} with a 'user' role authentication token in the Authorization header. Expect a 200 OK response, but confirm that only authorized resources or data for a regular user are included in the response. Verify that any restricted admin-only resources are not accessible to the user role. user:{account}.\n",
+                        # f"send a GET request to {endpoint} with a 'guest' role authentication token (or without an authentication token if guests are unauthenticated). Expect either a 403 Forbidden or a 404 Not Found response if guests are not permitted to access the endpoint, or a 200 OK with limited data if some resources are publicly accessible to guests."
+                    ],
+                    "path": [endpoint, ],
+                    "expected_response_code": [
+                        # "200 OK for admin, confirming full access.",
+                        "200 OK for users, confirming access is limited to non-admin resources.",
+                        # "403 Forbidden or 404 Not Found for guests if no access is granted, or 200 OK if limited access is applicable."
+                    ],
+                    "token": [  # self.admin.get("token"),
+                        token,
+                        # self.guest.get("token")
+                    ],
+                    "security": [
+                        "Verify that the response data includes complete access to the endpoint, reflecting the admin’s permissions.",
+                        "Verify that any restricted admin-only resources are not accessible to the user role.",
+                        "Verify that guest role has no or limited access."],
+
+                },
+                # Access Control to Specific Resources
+                {
+                    "objective": "Access Control to Specific Resources",
+                    "steps": [
+                        f"Accessing Resource with Owner's Credentials - Send a GET request to {endpoint} using an authentication token for User A (the owner of the resource). Expect a 200 OK response, confirming that User A can access their own data. Verify that the response includes the resource data as expected.user:{account}.\n",
+                        f"Attempting Access with Non-Owner's Credentials - Send a GET request to {endpoint_of_other_user} using an authentication token for User B (a different user who does not own the resource). Expect a 403 Forbidden or 404 Not Found response, indicating that access is denied to resources not owned by the authenticated user.user:{account}.\n",
+                        f"Attempting Access without Authentication (if applicable) - Send a GET request to {endpoint} without any authentication token. Expect a 401 Unauthorized response, confirming that unauthenticated access is restricted.user:{account}.\n"
+                    ],
+                    "path": [endpoint, endpoint_of_other_user, endpoint],
+                    "expected_response_code": [
+                        "200 OK when accessed by the owner, confirming correct owner access.",
+                        "403 Forbidden or 404 Not Found when accessed by a non-owner, confirming that access is appropriately restricted.",
+                        "401 Unauthorized for unauthenticated requests, ensuring that authentication is required for access."
+                    ],
+                    "token": [account.get("token"), self.create_random_bearer_token(), ""],
+                    "security": [
+                        "Ensure robust validation of user credentials and proper enforcement of access controls. Verify that the system correctly identifies ownership and prevents unauthorized access based on user authentication and authorization levels."
+                        ]
+                }
+
+                # This prompt tests the RBAC implementation by confirming that each user role has the appropriate access level:
+                # - Admin has full access
+                # - User has restricted access
+                # - Guest has no or limited access.
+            ]
+
+            endpoint = old_endpoint
+
+        return prompts
+
+    def verify_data_masking(self, endpoint, account, prompts):
+        prompts = prompts + [
+
+            {
+                "objective": "Verify Data Masking",
+                "steps": [
+                    f"Restricted Access Verification - Send a GET request to {endpoint} with a user who has restricted access, verifying that only permissible fields are included in the response.user:{account}.\n",
+                ],
+                "expected_response_code": [
+                    "200 OK for responses that correctly implement data masking, ensuring that data visibility is restricted according to user permissions."
+                ],
+                "path": [endpoint],
+                "token": [account.get("token")],
+                "security": [
+                    " Confirm that sensitive data is either masked or excluded, ensuring that only authorized fields are visible to the user. Ensure that the system effectively masks or excludes sensitive data for users with restricted access. Verify that data masking is applied correctly and consistently across all applicable endpoints to maintain confidentiality and compliance with data protection regulations."
+                ]
+            }
+
+        ]
+        return prompts
+
+    def test_create(self, endpoint, account, prompts, post_endpoint):
+        post_schema = post_endpoint.get("schema")
+        post_schema = self.fill_schema(post_schema)
+
+        prompts = prompts + [  # Check Permissions for CRUD Operations
+
+            # Create Operation:
+
+            {
+                "objective": "Check Permissions for CRUD Operations: Create",
+                "steps": [
+                    f"Authorized Creation - Send a POST request to {endpoint} with valid data to create a new resource like {post_schema}, using a user with creation permissions. Expect a 201 Created response if successful.",
+                    f"Unauthorized Creation - Then, attempt to send a similar request to {endpoint}  like {post_schema} with a user who lacks creation permissions, expecting a 403 Forbidden or similar error response."
+                ],
+                "path": [endpoint, endpoint],
+                "expected_response_code": [
+                    "201 Created for successful creation by authorized users, confirming that the system allows creation as expected.",
+                    "403 Forbidden for unauthorized creation attempts, confirming that the system properly restricts creation permissions."
+                ],
+                "token": [account.get("token"), account.get("token")],
+                "security": [
+                    "Ensure that the system robustly validates user permissions before processing CRUD operations. "
+                    "Confirm that only users with the appropriate permissions can create resources, and that unauthorized attempts are correctly identified and denied."]
+            }
+        ]
+
+        return prompts
+
+    def test_get(self, endpoint, account, prompts):
+
+        get_endpoint, endpoint_of_other_user = self.replace_ids(account, endpoint)
+
+        prompts = prompts + [
+
+            # Read Operation:
+
+            {
+                "objective": "Check Permissions for CRUD Operations: Read",
+                "steps": [
+                    f"Authorized Read - Send a GET request to {get_endpoint} with a user who has read permissions, verifying that the correct data is returned.",
+                    f"Unauthorized Read - Attempt the same request to {endpoint_of_other_user} with a user who lacks read permissions, and verify that the response returns a 403 Forbidden or 404 Not Found status."
+                ],
+                "path": [get_endpoint, get_endpoint],
+                "expected_response_code": [
+                    "200 OK for successful data retrieval by authorized users, confirming that the system allows reading of data as expected.",
+                    "403 Forbidden or 404 Not Found for unauthorized read attempts, confirming that the system properly restricts reading permissions."
+                ],
+                "token": [account.get("token"), account.get("token")],
+
+                "security": [
+                    "Ensure that the system robustly validates user permissions before allowing access to read operations. Confirm that only users with the appropriate permissions can access data, and that unauthorized attempts are correctly identified and denied, preventing data leaks."]
+            }]
+
+
+        return prompts
+
+    def test_put(self, put_endoint, account, prompts):
+        endpoint = put_endoint.get("path")
+        put_endoint_schema = put_endoint.get("schema")
+        put_endoint, endpoint_of_other_user = self.replace_ids(account, endpoint)
+        prompts = prompts + [
+
+            # Update Operation:
+
+            {
+                "objective": "Check Permissions for CRUD Operations: Update",
+                "steps": [
+                    f"Authorized Update - Send a PUT or PATCH request to {put_endoint} to update a resource using a user with update permissions. Expect the resource to be modified, indicated by a 200 OK or 204 No Content response schema:{put_endoint_schema}.",
+                    f"Unauthorized Update - Then, repeat the request with a user to {endpoint_of_other_user}who lacks update permissions, expecting a 403 Forbidden response."
+                ],
+                "path": [put_endoint, put_endoint],
+                "token": [account.get("token"), account.get("token")],
+
+                "expected_response_code": [
+                    "200 OK or 204 No Content for successful modification by authorized users, confirming that the system allows updates as expected.",
+                    "403 Forbidden for unauthorized update attempts, confirming that the system properly restricts updating permissions."
+                ],
+                "security": [
+                    "Ensure that the system robustly validates user permissions before allowing access to update operations. Confirm that only users with the appropriate permissions can modify resources, and that unauthorized attempts are correctly identified and denied, preventing unauthorized data modification."]}
+
+        ]
+        return prompts
+
+    def test_delete(self, endpoint, account, prompts):
+        delete_endpoint, endpoint_of_other_user = self.replace_ids(account, endpoint)
+
+        prompts = prompts + [
+
+            # Delete Operation:
+
+            {
+                "objective": "Check Permissions for CRUD Operations: Delete",
+                "steps": [
+                    f"Authorized Deletion - Send a DELETE request to {delete_endpoint} with a user who has delete permissions, verifying a successful deletion with a 200 OK or 204 No Content response.",
+                    f"Unauthorized Deletion - Then, attempt to delete the resource with a user {endpoint_of_other_user} who lacks delete permissions, expecting a 403 Forbidden response."
+                ],
+                "token": [account.get("token"), account.get("token")],
+                "path": [delete_endpoint, delete_endpoint],
+                "expected_response_code": [
+                    "200 OK or 204 No Content for successful deletion by authorized users, confirming that the system allows deletions as expected.",
+                    "403 Forbidden for unauthorized deletion attempts, confirming that the system properly restricts deleting permissions."
+                ],
+                "security": [
+                    "Ensure that the system robustly validates user permissions before allowing access to delete operations. Confirm that only users with the appropriate permissions can delete resources, and that unauthorized attempts are correctly identified and denied, preventing unauthorized resource removal."]
+            }]
+
+        return prompts
+
+    def test_access_scope(self, get_endpoint, post_endpoint, put_endpoint, account, prompts):
+        if isinstance(get_endpoint, dict):
+            get_endpoint = get_endpoint.get("path")
+        if isinstance(post_endpoint, dict):
+            post_endpoint = post_endpoint.get("path")
+        if isinstance(put_endpoint, dict):
+            put_endoint = put_endpoint.get("path")
+        if "api" in get_endpoint and post_endpoint and put_endpoint:
+            if account["api"] in get_endpoint and account["api"] in post_endpoint and account[
+                "api"] in put_endpoint:
+                id = account.get("id")
+                get_endpoint = get_endpoint.replace("{id}", str(account.get("id")))
+                post_endoint = post_endpoint.replace("{id}", str(account.get("id")))
+                put_endoint = put_endoint.replace("{id}", str(account.get("id")))
+                prompts = prompts + [
+
+                    # Access Token Scope Testing
+                    {
+                        "objective": "Access Token Scope Testing",
+                        "steps": [
+                            f"Testing Read-Only Scope for Data Retrieval - Send a GET request to {get_endpoint} using a read-only access token. Verify that the response status is 200 OK, allowing data retrieval.",
+                            f"Attempting Write Operation with Read-Only Token - Attempt to modify data on {put_endoint} by sending a POST, PUT, or DELETE request using the same read-only access token.",
+                            f"Testing Full-Access Token (if applicable) - Send a POST request to {post_endoint} using a full-access token to verify that write actions are permitted."
+                        ],
+                        "token": [account.get("token"), self.create_random_bearer_token(),
+                                  account.get("token")],
+                        "path": [get_endpoint, put_endoint, post_endoint],
+
+                        "expected_response_code": [
+                            "200 OK for successful data retrieval using a read-only token, confirming the enforcement of read-only access.",
+                            "403 Forbidden for attempted write operations with a read-only token, confirming that the token scope correctly restricts write actions.",
+                            "200 OK or 201 Created for successful write actions using a full-access token, confirming that full-access privileges are appropriately granted."
+                        ],
+                        "security": [
+                            "Ensure that the a A read-only access token permits data retrieval (GET request).",
+                            "The same read-only token denies access to write operations (POST, PUT, DELETE requests).",
+                            "A full-access token (if applicable) allows write actions, validating proper enforcement of token scopes."]
+                    }
+                ]
+
+        else:
+            if "id}" in get_endpoint:
+                get_endpoint = self.replace_placeholders_with_1(get_endpoint, account.get("id"))
+                prompts = prompts + [
+
+                    # Access Token Scope Testing
+                    {
+                        "objective": "Access Token Scope Testing",
+                        "steps": [
+                            f"Testing Read-Only Scope for Data Retrieval - Send a GET request to {get_endpoint} using a read-only access token. Verify that the response status is 200 OK, allowing data retrieval.",
+                            f"Attempting Write Operation with Read-Only Token - Attempt to modify data on {put_endpoint} by sending a POST, PUT, or DELETE request using the same read-only access token.",
+                            f"Testing Full-Access Token (if applicable) - Send a POST request to {post_endpoint} using a full-access token to verify that write actions are permitted."
+                        ],
+                        "token": [account.get("token"), self.create_random_bearer_token(),
+                                  account.get("token")],
+                        "path": [get_endpoint, put_endpoint, post_endpoint],
+
+                        "expected_response_code": [
+                            "200 OK for successful data retrieval using a read-only token, confirming the enforcement of read-only access.",
+                            "403 Forbidden for attempted write operations with a read-only token, confirming that the token scope correctly restricts write actions.",
+                            "200 OK or 201 Created for successful write actions using a full-access token, confirming that full-access privileges are appropriately granted."
+                        ],
+                        "security": [
+                            "Ensure that the a A read-only access token permits data retrieval (GET request).",
+                            "The same read-only token denies access to write operations (POST, PUT, DELETE requests).",
+                            "A full-access token (if applicable) allows write actions, validating proper enforcement of token scopes."]
+                    }
+                ]
+        return prompts
+
+    # This prompt tests that:
+    # - A read-only access token permits data retrieval (GET request).
+    # - The same read-only token denies access to write operations (POST, PUT, DELETE requests).
+    # - A full-access token (if applicable) allows write actions, validating proper enforcement of token scopes.
+
+    def test_crud(self, endpoints, prompts):
+        post_endpoints = self.get_correct_endpoints_for_method("protected_endpoint", "POST")
+        delete_endpoints = self.get_correct_endpoints_for_method("protected_endpoint", "DELETE")
+        put_endpoints = self.get_correct_endpoints_for_method("protected_endpoint", "PUT")
+
+        for account in self.accounts:
+
+            if "id" in account.keys():
+                id = account.get("id")
+            else:
+                id = 1
+
+            for post_endpoint in post_endpoints:
+
+                if "api" in post_endpoint and len(post_endpoint.split("/")) > 0:
+                    if account["api"] in post_endpoint:
+                        endpoint = post_endpoint.replace("{id}", str(account.get("id")))
+                        prompts = self.test_create(endpoint, account, prompts, post_endpoint=post_endpoint)
+
+                    else:
+                        prompts = self.test_create(post_endpoint.get("path"), account, prompts,
+                                                   post_endpoint=post_endpoint)
+                else:
+                    prompts = self.test_create(post_endpoint.get("path"), account, prompts, post_endpoint)
+
+            for get_endpoint in endpoints:
+                if isinstance(get_endpoint, dict):
+                    get_endpoint = get_endpoint.get("path")
+
+                if "api" in get_endpoint and "id" in account.keys():
+                    if account["api"] in get_endpoint and isinstance(account["id"], int):
+
+                        prompts = self.test_get(get_endpoint, account, prompts)
+
+                    else:
+                        prompts = self.test_get(get_endpoint, account, prompts)
+
+                else:
+                    if "{id}" in get_endpoint:
+
+                        id = 1
+                        if "OWASP" in self.config.get("name") and "basket" not in get_endpoint:
+                            continue
+
+                        prompts = self.test_get(get_endpoint, account, prompts)
+
+            for put_endoint in put_endpoints:
+                if "api" in put_endoint:
+                    if account["api"] in put_endoint:
+                        prompts = self.test_put(put_endoint, account, prompts)
+                else:
+
+                    prompts = self.test_put(put_endoint, account, prompts)
+
+            for delete_endpoint in delete_endpoints:
+                id = 2
+                if isinstance(delete_endpoint, dict):
+                    delete_endpoint_schema = delete_endpoint.get("schema")
+                    delete_endpoint = delete_endpoint.get("path")
+                if "api" in delete_endpoint and "id" in delete_endpoint:
+                    prompts = self.test_delete(delete_endpoint, account, prompts)
+                else:
+                    prompts = self.test_delete(delete_endpoint, account, prompts)
+
+            for get_endpoint in endpoints:
+                for post_endpoint in post_endpoints:
+                    for put_endpoint in put_endpoints:
+                        prompts = self.test_access_scope(get_endpoint, post_endpoint, put_endpoint, account, prompts)
+
+        return prompts
+
+    def test_roles(self, post_endpoint, schema, account, prompts):
+        prompts.append(
+            # Test for Manipulation of User Roles
+            {
+                "objective": "Test for Manipulation of User Roles",
+                "steps": [
+                    f"Send a POST request to {post_endpoint} based on schema {schema} attempting to elevate a standard user's role to an admin.",
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "403 Forbidden if the API correctly identifies and prevents role escalation attempts.",
+                    "400 Bad Request if the request format or parameters are invalid."
+                ],
+                "security": [
+                    "Verify if the API prevents unauthorized role manipulations.",
+                    "Ensure that the API enforces strict role-based access controls and prevents unauthorized role changes.",
+                    "Verify that unauthorized role manipulation attempts are logged and do not expose any vulnerabilities."
+                ]
+            }
+
+            # This prompt tests if the API prevents unauthorized role escalation.
+        )
+        return prompts
+
