+++ src/hackingBuddyGPT/utils/prompt_generation/information/_pentesting_core.py
@@ -1,0 +1,816 @@
+# This file is part of the PenTestingInformation split (see pentesting_information.py).
+import base64
+import copy
+import csv
+import glob
+import os
+import random
+import re
+import secrets
+from typing import List
+
+from hackingBuddyGPT.utils.openapi.openapi_parser import OpenAPISpecificationParser
+from hackingBuddyGPT.utils.prompt_generation.information.prompt_information import (
+    PromptPurpose,
+)
+from faker import Faker
+
+
+class PenTestingCore:
+    """Core setup, step dispatch and shared endpoint/schema/account helpers."""
+
+    def __init__(self, openapi_spec_parser: OpenAPISpecificationParser, config) -> None:
+        """
+        Initializes the PenTestingInformation with optional authentication credentials.
+
+        Args:
+            openapi_spec_parser (OpenAPISpecificationParser): An instance of OpenAPISpecificationParser.
+            username (str, optional): Username for authentication, if necessary. Defaults to an empty string.
+            password (str, optional): Password for authentication, if necessary. Defaults to an empty string.
+        """
+
+        # Set basic authentication details
+        if "admin" in config:
+            self.admin = config["admin"]
+        else:
+            self.admin = None
+        self.guest = None
+        self.credentials = {}
+        self.resources = {}
+        self.valid_token = None
+        self.current_post_endpoint = None  # TODO
+        self.faker = Faker()
+        self.username = self.faker.email().lower()
+        self.password = self.faker.password()
+        self.available_numbers = []
+        self.config = config
+        file = self.get_file(self.config.get("csv_file"))
+        self._credentials = []
+        if file != "Not found":
+            with open(file[0]) as csvfile:
+                reader = csv.reader(csvfile, delimiter=',')
+                for row in reader:
+                    self._credentials.append({
+                        'username': row[0],
+                        'password': row[1]
+                })
+
+        # Parse endpoints and their categorization from the given parser instance
+        categorized_endpoints = openapi_spec_parser.classify_endpoints(self.config.get("name"))
+
+        # Assign schema and endpoint attributes directly from the parser methods
+        self.schemas = openapi_spec_parser.get_schemas()
+        self.endpoints = openapi_spec_parser.get_endpoints()
+        self.openapi_spec_parser = openapi_spec_parser
+        self.get_comment_ep()
+
+        # Assign categorized endpoint types to attributes
+        self.assign_endpoint_categories(categorized_endpoints)
+        self.accounts = []
+        opt_endpoints = [ep for ep in self.endpoints if "otp" in ep]
+
+        self.brute_force_accounts = []
+        if self.admin is not None:
+            admin = self.config.get("admin").get("email")
+            self.assign_brute_force_endpoints(admin)
+
+        self.pentesting_step_list = [PromptPurpose.SETUP,
+                                     PromptPurpose.VERIY_SETUP,
+                                     PromptPurpose.AUTHENTICATION,
+                                     PromptPurpose.AUTHORIZATION,
+                                     PromptPurpose.SPECIAL_AUTHENTICATION,
+                                     PromptPurpose.INPUT_VALIDATION,
+                                     PromptPurpose.ERROR_HANDLING_INFORMATION_LEAKAGE,
+                                     PromptPurpose.SESSION_MANAGEMENT,
+                                     PromptPurpose.CROSS_SITE_SCRIPTING,
+                                     PromptPurpose.CROSS_SITE_FORGERY,
+                                     PromptPurpose.BUSINESS_LOGIC_VULNERABILITIES,
+                                     PromptPurpose.RATE_LIMITING_THROTTLING,
+                                     PromptPurpose.SECURITY_MISCONFIGURATIONS,
+                                     PromptPurpose.LOGGING_MONITORING
+                                     ]
+
+    def assign_endpoint_categories(self, categorized_endpoints):
+        """
+        Assign categorized endpoint types to instance attributes from given categorized endpoints dictionary.
+
+        Args:
+            categorized_endpoints (dict): A dictionary containing categorized endpoints.
+        """
+        self.resource_intensive_endpoint = categorized_endpoints.get('resource_intensive_endpoint')
+
+        self.secure_action_endpoint = categorized_endpoints.get('secure_action_endpoint')
+        self.role_access_endpoint = categorized_endpoints.get('role_access_endpoint')
+        self.sensitive_data_endpoint = categorized_endpoints.get('sensitive_data_endpoint')
+        self.sensitive_action_endpoint = categorized_endpoints.get('sensitive_action_endpoint')
+        self.login_endpoint = categorized_endpoints.get('login_endpoint')
+        self.account_endpoint = categorized_endpoints.get('account_creation')
+        self.auth_endpoint = categorized_endpoints.get('auth_endpoint')
+        self.generate_iter_and_assign_current_endpoints(categorized_endpoints)
+        self.analysis_step_list = [PromptPurpose.ANALYSIS, PromptPurpose.DOCUMENTATION,
+                                   PromptPurpose.REPORTING]
+        self.categorized_endpoints = categorized_endpoints
+        self.tokens = {}
+        self.counter = 0
+
+    def generate_iter_and_assign_current_endpoints(self, categorized_endpoints):
+        for key in ['public_endpoint', 'protected_endpoint', 'refresh_endpoint']:
+            endpoint_list = categorized_endpoints.get(key, [])
+            if endpoint_list:
+                setattr(self, f"{key}_iterator", iter(endpoint_list))
+                setattr(self, f"current_{key}", next(getattr(self, f"{key}_iterator"), None))
+            else:
+                setattr(self, f"{key}_iterator", iter([]))
+                setattr(self, f"current_{key}", None)
+
+    def explore_steps(self, purpose: PromptPurpose) -> List[str]:
+        """
+        Provides initial penetration testing steps for the given purpose.
+
+        Args:
+            purpose (PromptPurpose): The purpose for which testing steps are required.
+
+        Returns:
+            list: A list of steps corresponding to the specified purpose.
+        """
+        # Map purposes to their corresponding methods
+        purpose_methods = {
+            PromptPurpose.SETUP: self.setup_test,
+            PromptPurpose.VERIY_SETUP: self.verify_setup,
+            PromptPurpose.AUTHENTICATION: self.generate_authentication_prompts,
+            PromptPurpose.AUTHORIZATION: self.generate_authorization_prompts,
+            PromptPurpose.SPECIAL_AUTHENTICATION: self.generate_special_authentication,
+            PromptPurpose.INPUT_VALIDATION: self.generate_input_validation_prompts,
+            PromptPurpose.ERROR_HANDLING_INFORMATION_LEAKAGE: self.generate_error_handling_prompts,
+            PromptPurpose.SESSION_MANAGEMENT: self.generate_session_management_prompts,
+            PromptPurpose.CROSS_SITE_SCRIPTING: self.generate_xss_prompts,
+            PromptPurpose.CROSS_SITE_FORGERY: self.generate_csrf_prompts,
+            PromptPurpose.BUSINESS_LOGIC_VULNERABILITIES: self.generate_business_logic_vul_prompts,
+            PromptPurpose.RATE_LIMITING_THROTTLING: self.generate_rate_limit_throttling,
+            PromptPurpose.SECURITY_MISCONFIGURATIONS: self.generate_security_misconfiguration_prompts,
+            PromptPurpose.LOGGING_MONITORING: self.generate_logging_monitoring_prompts
+        }
+
+        # Call the appropriate method based on the purpose
+        if purpose in purpose_methods:
+            return purpose_methods[purpose]()
+        else:
+            raise ValueError(f"Invalid purpose: {purpose}")
+
+    def get_analysis_step(self, purpose: PromptPurpose = None, response: str = "", additional_context: str = "") -> str:
+        """
+        Provides prompts for analysis based on the provided response for various purposes using an LLM.
+
+        Args:
+            response (str, optional): The HTTP response to analyze. Default is an empty string.
+
+        Returns:
+            dict: A dictionary where each key is a PromptPurpose and each value is a list of prompts.
+        """
+        if purpose == PromptPurpose.ANALYSIS:
+            return f"Given the following parsed HTTP response:\n{response}\n" \
+                   f"Based on this context: {additional_context}\n" \
+                   "Analyze this response to determine in form of a RecordNote:\n" \
+                   "1. Whether the status code is appropriate for this type of request.\n" \
+                   "2. If the headers indicate proper security and rate-limiting practices. \n" \
+                   "3. If the headers include sensitive information like 'Server', 'X-Powered', 'X-Frame-Options', 'Cache-Control', 'Strict-Transport-Security', 'Set-Cookie', 'X-Request-ID', 'Accept-Encoding', 'Referer', and 'X-API-Version' ALso add why this can cause a vulnerability.\n" \
+                   "4. Whether the response body is correctly handled."
+            # "Keep your analysis short."
+
+        if purpose == PromptPurpose.DOCUMENTATION:
+            return f"Based on the analysis provided, document the findings in form of a RecordNote:\n{response}."
+            # f" Keep your analysis short."
+
+        if purpose == PromptPurpose.REPORTING:
+            return (f"Based on the documented findings : {response}.\n"
+                    f"Suggest any improvements or issues that should be reported based on the findings to the API developers in form of a RecordNote.")
+            # f"Keep your analysis short."
+
+    def get_steps_of_phase(self, purpose):
+        steps = self.explore_steps(purpose)
+        return steps
+
+    def next_testing_endpoint(self):
+        self.current_public_endpoint = next(self.public_endpoint_iterator, None)
+        self.current_protected_endpoint = next(self.protected_endpoint_iterator, None)
+        self.current_refresh_endpoint = next(self.refresh_endpoint_iterator, None)
+
+    def setup_test(self) -> List:
+        prompts = []
+        counter = 0
+        post_account = self.get_correct_endpoints_for_method("account_creation", "POST")
+        prompts, counter = self.generate_user(post_account, counter, prompts)
+        if len(self.accounts) == 1:  # ensure that there are at least two users
+            prompts, counter = self.generate_user(post_account, 1, prompts)
+
+        return prompts
+
+    def verify_setup(self) -> List:
+        prompts = []
+
+        get_account = self.get_correct_endpoints_for_method("public_endpoint",
+                                                            "GET") + self.get_correct_endpoints_for_method(
+            "protected_endpoint", "GET")
+
+        get_account = [ep for ep in get_account if ep.get("path").endswith("user") or ep.get("path").endswith("login")]
+        if len(get_account) == 0:
+            get_account = [
+                self.endpoints[path]for path, methods in self.endpoints.items()
+                if path.endswith("/users/{id}") and "get" in methods
+            ]
+
+        for acc in get_account:
+            for account in self.accounts:
+                account_path = acc.get("path")
+                account_schema = acc.get("schema")
+                if "api" in account_path:
+                    if account["api"] in account_path:
+                        if "user" and "id" in account_path:
+                            account_path = account_path.replace("{id}", str(account.get("id")))
+                        prompts = prompts + [{
+                            "objective": "Check if user was created",
+                            "steps": [
+                                f"Endpoint to use : {account_path}\n"
+                                f"Send a GET request to the {account_path} with the with the correct schema {account_schema} with user:{account}.\n"
+                            ],
+                            "path": [account_path],
+                            "token": [account.get("token")],
+                            "expected_response_code": ["200 OK", "201 Created"],
+                            "security": [
+                                f"Ensure that the returned user matches this user {account}"]
+                        }]
+                else:
+                    if "id}" in account_path:
+
+                        if isinstance(account.get("example"), dict):
+                            if "example" in account.keys():
+                                if "id" in account.get("example").keys():
+                                    account_path = account_path.replace("{id}",
+                                                                        str(account_schema.get("example").get("id")))
+                                else:
+                                    account_path = account_path.replace("{id}", str(account_schema.get("example")))
+                        else:
+                            account_path = self.replace_placeholders_with_1(account_path, account.get("id"))
+
+                    if account_schema:
+                        if "Authorization-Token" in account_schema.values():
+                            if "example" in account.keys() and "id" in account.get("example") and account.get(
+                                    "example").get("id") not in self.tokens.keys():
+                                description = account_schema.get("description") \
+                                    .replace("username", account.get("example").get("username")) \
+                                    .replace("password", account.get("example").get("password"))
+                                account_schema["description"] = description
+                                credentials = account.get("example").get("username") + ":" + account.get("example").get(
+                                    "password")
+                                self.tokens[account.get("example").get("id")] = base64.b64encode(
+                                    credentials.encode('utf-8')).decode('utf-8')
+
+                    prompts = prompts + [{
+                        "objective": "Check if user was created",
+                        "steps": [
+                            f"Endpoint to use : {account_path}\n"
+                            f"Send a GET request to the {account_path} with the with the correct schema {account_schema} with user:{account}.\n"
+                        ],
+                        "path": [account_path],
+                        "token": [account.get("token")],
+                        "expected_response_code": ["200 OK", "201 Created"],
+                        "security": [
+                            f"Ensure that the returned user matches this user {account}"]
+                    }]
+        for login in self.login_endpoint:
+            for account in self.accounts:
+
+                login_path, login_schema = self.get_path_and_schema(login)
+                if login_schema is None:
+                    continue
+                login_schema = self.fill_in_schema_with_account(login_schema, account)
+
+                if "api" in account.keys():
+                    if account["api"] in login_path:
+                        prompts = prompts + [
+                            {
+                                "objective": "Get Valid token",
+                                "steps": [
+                                    f"Endpoint to use : {login_path}\n"
+                                    f"Send a POST request to the {login_path} with the correct credentials of user:{account}.\n"
+                                    f"Request body should be in application/json and look similar to this: {{ {login_schema}}}"
+                                ],
+                                "path": [login_path],
+                                "token": [account.get("token")],
+                                "expected_response_code": ["200 OK", "201 Created"],
+                                "security": [
+                                    "Ensure that the correct expected code is received  confirming that access is restricted and no sensitive data is included in the response body."]
+                            }
+                        ]
+                else:
+
+                    if account is None:
+                        account = self.create_account(login_schema, login_path)
+                        if account is None:
+                            continue
+
+                    prompts = prompts + [
+                        {
+                            "objective": "Get Valid token",
+                            "steps": [
+                                f"Endpoint to use : {login_path}\n"
+                                f"Send a POST request to the {login_schema} with the correct credentials of user:{account}.\n"
+                                f"Request body should be in application/json and look similar to this: {{ {account}}}"
+                            ],
+                            "path": [login_path],
+                            "token": [""],
+                            "expected_response_code": ["200 OK", "201 Created"],
+                            "security": [
+                                "Ensure that the correct expected code is received  confirming that access is restricted and no sensitive data is included in the response body."]
+                        }
+                    ]
+
+        account = None
+        if len(prompts) == 0:
+            for login in self.login_endpoint:
+                login_path, login_schema = self.get_path_and_schema(login)
+                if login_schema is None:
+                    continue
+                if account is None:
+                    account = self.create_account(login_schema, login_path)
+                    if account is None:
+                        continue
+
+                prompts = prompts + [
+                    {
+                        "objective": "Get Valid token",
+                        "steps": [
+                            f"Endpoint to use : {login_path}\n"
+                            f"Send a POST request to the {login_schema} with the correct credentials of user:{account}.\n"
+                            f"Request body should be in application/json and look similar to this: {{ {login_schema}}}"
+                        ],
+                        "path": [login_path],
+                        "token": [""],
+                        "expected_response_code": ["200 OK", "201 Created"],
+                        "security": [
+                            "Ensure that the correct expected code is received  confirming that access is restricted and no sensitive data is included in the response body."]
+                    }
+                ]
+
+        return prompts
+
+    def replace_placeholders_with_1(self, path: str, id) -> str:
+        """
+        Replaces any curly-brace placeholders (e.g., '{videoid}', '{orderid}', '{someid}')
+        with the number '1' in the given path.
+
+        Example:
+            "/identity/api/v2/user/videos/{videoid}" -> "/identity/api/v2/user/videos/1"
+            "/workshop/api/shop/orders/{orderid}"    -> "/workshop/api/shop/orders/1"
+            "{somethingid}"                          -> "1"
+        """
+
+        def substitute(match):
+            # Extract the placeholder from the match
+            placeholder = match.group(0).strip('{}')
+            # Return the replacement for the placeholder if it exists
+            return id.get(placeholder, match.group(0))
+
+            # Regex to match anything in curly braces
+
+        # Regex to match anything in curly braces, e.g. {videoid}, {postid}, etc.
+
+        if id is None:
+            return path
+        if isinstance(id, int):
+            id = str(id)
+
+        return re.sub(r"\{[^}]+\}", id, path)
+
+    def key_in_path(self, path, dictionary):
+        # Split the path into components
+        path_parts = path.split('/')
+
+        # Check each part of the path to see if it is a key in the dictionary
+        for part in path_parts:
+            part = part.replace("{", "")
+            part = part.replace("}", "")
+
+            if part in dictionary:
+                return True, part  # Return True and the matching part if found
+
+        return False, None  # Return False and None if no part matches
+
+    def get_correct_endpoints_for_method(self, type_of_endpoint, method):
+        endpoints = []
+        for type_ep in self.categorized_endpoints.keys():
+            if type_of_endpoint == type_ep:
+                x = self.categorized_endpoints[type_of_endpoint]
+                if x is not None:
+                    for entry in x:  # Assuming x is a list of dictionaries
+                        if entry.get('method') == method:
+                            endpoints.append(entry)
+        return endpoints
+
+    def generate_random_numbers(self, length=10):
+
+        number = ''.join(str(random.randint(0, 9)) for _ in range(length))
+        while number in self.available_numbers:
+            number = ''.join(str(random.randint(0, 9)) for _ in range(length))
+
+        self.available_numbers.append(number)
+        return number
+
+    def get_credentials(self, schema, endpoint, new_user=False):
+        """
+            Fill username and password fields in the provided schema.
+
+            Args:
+                schema (dict): A schema dictionary containing an example.
+                username (str): The username to populate in the example.
+                password (str): The password to populate in the example.
+
+            Returns:
+                dict: Updated schema with username and password fields filled.
+            """
+        # Deep copy the schema to avoid modifying the original
+        updated_schema = copy.deepcopy(schema)
+        example = None
+
+        if schema is not None:
+            if "example" in updated_schema.keys():
+                updated_schema["example"] = self.fill_schema(updated_schema["example"])
+            else:
+                updated_schema = self.adjust_schema_with_examples(updated_schema)
+
+            if "example" in updated_schema:
+                example = updated_schema["example"]
+            if endpoint not in self.credentials.keys() or new_user:
+
+                # Check if 'example' exists and is a dictionary
+                if updated_schema is not None and "example" in updated_schema.keys():
+                    example = updated_schema.get("example")
+
+                if example is None:
+                    example = {}
+                    if "email" not in example or example["email"].startswith("{{"):
+                        example['email'] = self.faker.email()
+                    if "name" not in example or example["name"].startswith("{{"):
+                        example["name"] = self.faker.name()
+                    if "number" not in example:
+                        if schema is not None and "properties" in schema.keys():
+                            example["number"] = int(self.generate_random_numbers())
+                        else:
+                            example["number"] = 1
+                    if "username" in example and example["username"].startswith("{{"):
+                        example["username"] = self.faker.user_name()
+                else:
+                    if "email" in example and "{{" in example["email"]:
+                        example["email"] = self.faker.email()
+                    if "password" in example and "{{" in example["password"]:
+                        password = self.faker.password(special_chars=False)
+                        if "passwordRepeat" in example  and "{{" in example["passwordRepeat"]:
+                            example["passwordRepeat"] = password
+                        example["password"] = password
+                    if "number" in example:
+                        if "{{" in example["number"] or "phone" in example["number"]:
+                            example["number"] = int(self.generate_random_numbers())
+                    if "username" in example:
+                        example["username"] = self.faker.user_name()
+
+                if updated_schema is None:
+                    updated_schema = {}
+                updated_schema["example"] = example
+                self.credentials[endpoint] = updated_schema
+
+            else:
+                updated_schema = self.credentials[endpoint]
+
+        return updated_schema
+
+    def fill_schema(self, schema, params=None):
+        if params:
+            field_to_faker = params
+        else:
+
+            field_to_faker = {
+                'name': self.faker.name,
+                'email': self.faker.email,
+                'phone': self.faker.phone_number,
+                'password': self.faker.password,
+                'address': self.faker.address,
+                'city': self.faker.city,
+                'username': self.faker.user_name,
+                "old_email": "adam007@example.com",
+                "new_email": self.faker.email,
+                "price": -2000,
+                "number_of_repeats": 10000,
+            }
+        filled_schema = {}
+        if schema:
+            for key, value in schema.items():
+                # Attempt to find a Faker provider for the key
+
+                provider = field_to_faker.get(key)
+                if provider:
+                    # If a provider is found, use it to generate fake data
+                    if key == "password":
+                        filled_schema[key] = self.faker.password(special_chars=False)
+                    else:
+                        if not callable(provider):
+                            filled_schema[key] = self.faker.random_letters()
+                        else:
+                            filled_schema[key] = provider()
+                else:
+                    # If no provider is found, revert to a default or keep the original value
+                    filled_schema[key] = value
+        return filled_schema
+
+    def adjust_schema_with_examples(self, schema: dict) -> dict:
+        """
+        Move 'example' values from each property into a separate 'example' dict.
+        """
+        new_schema = schema.copy()
+        example_dict = {}
+
+        if 'properties' in schema:
+            for field, field_props in schema['properties'].items():
+                if 'example' in field_props:
+                    example_dict[field] = field_props.pop('example')
+                if "properties" in field_props:
+                    if field not in example_dict or not isinstance(example_dict[field], dict):
+                        example_dict[field] = {}
+
+                    for field1, field_props1 in field_props['properties'].items():
+                        if 'example' in field_props1:
+                            example_dict[field][field1] = field_props1.pop('example')
+
+
+            # Add collected examples
+            if example_dict:
+                new_schema['example'] = example_dict
+
+        return new_schema
+
+    def create_random_bearer_token(self, length=16):
+        """
+        Generates a random token using hex encoding and prefixes it with "Bearer ".
+        :param length: Number of bytes for the random token (each byte becomes two hex characters).
+        :return: A string in the format "Bearer <random-hex>".
+        """
+        token_value = secrets.token_hex(length)
+        return f"{token_value}"
+
+    def get_invalid_credentials(self, account):
+        invalid_account = {}
+        for values, keys in account.items():
+            if isinstance(values, str):
+                invalid_account[keys] = values + "1"
+            elif values.isnumeric():
+                invalid_account[keys] = values + 1
+            else:
+                invalid_account[keys] = "_" + values
+        return invalid_account
+
+    def create_account(self, login_schema, login_path):
+        account = self.get_credentials(login_schema, login_path, new_user=True)
+        if account is None:
+            return None
+        if "example" in account:
+            account = account.get("example")
+        account["x"] = self.counter
+        parts = [api for api in login_path.split("/") if api.strip()]
+        api = [part for part in parts if part.startswith("api")]
+        if len(api) > 0:
+            api = api[0]
+            account["api"] = api
+        return account
+
+    def assign_brute_force_endpoints(self, admin):
+        file = self.config.get("password_file")
+        if file == "Not found":
+            return
+        else:
+            password_list_path = self.get_file(file)[0]
+
+        # Open the password list file
+        with open(password_list_path, "r") as file:
+            passwords = file.readlines()
+
+        # Strip any extra whitespace characters (newlines, spaces)
+        passwords = [password.strip() for password in passwords]
+
+        # Start brute-force attack
+        for password in passwords:
+            # Create the data for the POST request
+            data = {
+                'username': admin,
+                'password': password
+            }
+            self.brute_force_accounts.append(data)
+
+    def get_comment_ep(self):
+        for ep in self.endpoints:
+            if "comment" in ep:
+                return ep
+
+    def fill_in_schema_with_account(self, login_schema, account):
+        if login_schema is not None:
+            if "token" in login_schema and "token" in account:
+                login_schema["token"] = account["token"]
+            if "email" in login_schema and "email" in account:
+                login_schema["email"] = account["email"]
+            if "password" in login_schema and "password" in account:
+                login_schema["password"] = account["password"]
+            if "username" in login_schema and "username" in account:
+                login_schema["username"] = account["username"]
+        return login_schema
+
+    def resource_prompts(self, endpoint, account, prompts):
+        key_found, key = self.key_in_path(endpoint, self.resources)
+        if key_found:
+            for resource in self.resources[key]:
+                new_endpoint = endpoint.replace("{", "").replace("}", "").replace(key, resource)
+
+                prompts.append(
+
+                    # Access Control to Specific Resources
+                    {
+                        "objective": "Access Control to Specific Resources",
+                        "steps": [
+                            f"Accessing Resource with Owner's Credentials - Send a GET request to {endpoint} using an authentication token for User A (a different user who does not own the resource). user:{account}.\n",
+                            f"Attempting Access with Non-Owner's Credentials - Send a GET request to {endpoint} using an authentication token for User B (a different user who does not own the resource). user:{account}.\n",
+                            f"Attempting Access without Authentication (if applicable) - Send a GET request to {endpoint} without any authentication token. Expect a 401 Unauthorized response, confirming that unauthenticated access is restricted."
+                        ],
+                        "path": [new_endpoint, new_endpoint, new_endpoint],
+                        "expected_response_code": [
+                            "403 Forbidden or 404 Not Found when accessed by a non-owner, confirming that access is appropriately restricted.",
+                            "403 Forbidden or 404 Not Found when accessed by a non-owner, confirming that access is appropriately restricted.",
+                            "401 Unauthorized for unauthenticated requests, ensuring that authentication is required for access."
+                        ],
+                        "token": [account.get("token"), self.create_random_bearer_token(), ""],
+                        "security": [
+                            "Ensure robust validation of user credentials and proper enforcement of access controls. Verify that the system correctly identifies ownership and prevents unauthorized access based on user authentication and authorization levels."
+                            ]}
+
+                    # This prompt tests access control for resources that are user-owned by verifying:
+                    # - Only the resource owner can access the data (User A)
+                    # - Non-owners receive a restricted response (User B)
+                    # - Unauthenticated requests are denied access (if applicable)
+                )
+        return prompts
+
+    def resource_endpoints(self, prompts):
+        resource_endpoints = []
+        for ep in self.endpoints:
+            if "_id" not in ep and ("posts" and "recent" in ep or "mechanic" and "workshop" in ep or "dashboard" in ep):
+                resource_endpoints.append(ep)
+
+        resource_endpoints = list(set(resource_endpoints))
+
+        for resource_endpoint in resource_endpoints:
+            for account in self.accounts:
+                token = account.get("token")
+                prompts = prompts + [
+                    {
+                        "objective": "Get other data",
+                        "steps": [
+                            f"Send a GET request to the {resource_endpoint} with the token {token}.\n"
+                        ],
+                        "path": [resource_endpoint],
+                        "token": [account.get("token")],
+                        "expected_response_code": ["200 OK", "201 Created"],
+                        "security": [
+                            "Ensure that the correct expected code is received  confirming that access is restricted and no sensitive data is included in the response body."]
+                    }
+
+                ]
+        return prompts
+
+    def replace_id_placeholder(self, path: str, value: str) -> str:
+        return re.sub(r'\{[a-zA-Z0-9_]*_id\}', value, path)
+
+    def generate_user(self, post_account, counter, prompts):
+        for account in post_account:
+            account_path = account.get("path")
+            account_schema = account.get("schema")
+            if self.config.get("name") == "crapi":
+                account_user = self.create_account(login_schema=account_schema, login_path=account_path)
+
+
+            else:
+                account_user = self.get_credentials(account_schema, account_path, new_user=True).get("example")
+            if account_user is None:
+                continue
+            account_user["x"] = counter
+            if "api" in account_path:
+                parts = [api for api in account_path.split("/") if api.strip()]
+                api = [part for part in parts if part.startswith("api")]
+                api = api[0]
+                account_user["api"] = api
+            if self.config.get("name") == "vAPI":
+                text = f'{account_user.get("username")}:{account_user.get("password")}'
+
+                account_user["token"] = base64.b64encode(text.encode()).decode()
+
+            prompt = {
+                "objective": "Setup tests",
+                "steps": [
+                    f"Create an account by sending a POST HTTP request to the correct endpoint from this {account_path} with these credentials of user:{account_user}.\n"
+                    f"Request body should be in application/json and look similar to this: {{ {account_user}}}"],
+                "expected_response_code": ["200 OK", "201 Created"],
+                "token": [""],
+                "path": [account_path],
+                "security": [
+                    "Ensure that the correct expected code is received  confirming that access is restricted and no sensitive data is included in the response body."]
+            }
+
+            self.accounts.append(account_user)
+            prompts = prompts + [prompt]
+            counter = counter + 1
+
+        return prompts, counter
+
+    def replace_ids(self, account, endpoint, given_id=None):
+
+        if given_id is None:
+            id = account.get("id", 1)
+        else:
+            id = given_id
+        other_id = self.get_other_id(id, account)
+
+        new_endpoint = endpoint.replace("{id}", str(id))
+        endpoint_of_other_user = endpoint.replace("{id}", str(other_id))
+
+        # Handle {id}
+        if "{id}" in endpoint:
+            if "example" in account and "id" in account["example"]:
+                id = account["example"]["id"]
+                other_id = id - 1 if account == self.accounts[-1] else id + 1
+                new_endpoint = endpoint.replace("{id}", str(id))
+                endpoint_of_other_user = endpoint.replace("{id}", str(other_id))
+            else:
+
+                new_endpoint = endpoint.replace("{id}", str(id))
+                endpoint_of_other_user = endpoint.replace("{id}", str(other_id))
+        # Handle _id mostly for resources
+        elif "_id}" in endpoint:
+            key_found, key = self.key_in_path(endpoint, self.resources)
+            if key_found == True and key is not None:
+                key = str(key)
+                first_id = self.resources[key][0]
+                if len(self.resources[key]) > 1:
+                    second_id = random.choice(self.resources[key][1:])
+                else:
+                    second_id = 1  # fallback to same id if no other id available
+                new_endpoint = endpoint.replace("{", "").replace("}", "").replace(key, first_id)
+                endpoint_of_other_user = endpoint.replace("{", "").replace("}", "").replace(key, second_id)
+            else:
+                other_id = self.get_other_id(id, account)
+                new_endpoint = self.replace_id_placeholder(endpoint, str(id))
+                endpoint_of_other_user = self.replace_id_placeholder(endpoint, str(other_id))
+
+            if given_id is not None:
+                other_id = self.get_other_id(id, account)
+                new_endpoint = self.replace_id_placeholder(endpoint, str(given_id))
+                endpoint_of_other_user = self.replace_id_placeholder(endpoint, str(other_id))
+
+        return new_endpoint, endpoint_of_other_user
+
+    def get_other_id(self, id, account):
+        if str(id).isdigit():
+
+            other_id = id - 1 if account == self.accounts[-1] else id + 1
+        else:
+            current_index = self.accounts.index(account)
+
+            # Pick next account if not last, else pick previous
+            other_account = self.accounts[current_index + 1] if current_index < len(self.accounts) - 1 else \
+                self.accounts[current_index - 1]
+
+            other_id = other_account.get("id", 1)
+        if other_id is None:
+            other_id = 2
+
+
+        return other_id
+
+    def get_file(self, param):
+        # Get current file directory
+        current_dir = os.path.dirname(__file__)
+
+        # Go up one level
+        parent_dir = os.path.abspath(os.path.join(current_dir, ".."))
+        parent_dir = parent_dir.split("/src")[0]
+
+        # Search for file (glob is recursive-friendly)
+        file = glob.glob(os.path.join(parent_dir, param), recursive=True)
+
+        if not file or param == "":
+            return "Not found"
+        return file
+
+    def get_path_and_schema(self, login):
+        login_path = login.get("path")
+        login_schema = login.get("schema")
+        if "example" not in login_schema:
+            login_schema = self.adjust_schema_with_examples(login_schema)
+        login_schema = login_schema.get("example")
+
+        return login_path, login_schema
+
