+++ src/hackingBuddyGPT/utils/prompt_generation/information/_pentesting_injection.py
@@ -1,0 +1,795 @@
+# This file is part of the PenTestingInformation split (see pentesting_information.py).
+import base64
+import copy
+import csv
+import glob
+import os
+import random
+import re
+import secrets
+from typing import List
+
+from hackingBuddyGPT.utils.openapi.openapi_parser import OpenAPISpecificationParser
+from hackingBuddyGPT.utils.prompt_generation.information.prompt_information import (
+    PromptPurpose,
+)
+from faker import Faker
+
+
+class InjectionScenarios:
+    """InjectionScenarios: scenario/helper methods mixed into PenTestingInformation."""
+
+    def generate_input_validation_prompts(self):
+        """
+           Generate prompts for testing input validation vulnerabilities on POST endpoints.
+
+           This method targets both protected and public POST endpoints and performs:
+           - SQL Injection testing using the account context and endpoint schema.
+           - General input validation testing (e.g., missing fields, invalid types).
+
+           For each endpoint and account combination, the method replaces path parameters (like {id})
+           and checks the relevant OpenAPI schema to craft test cases.
+
+           Returns:
+               list: A list of prompts designed to evaluate input validation robustness.
+           """
+        prompts = []
+
+        endpoints = self.get_correct_endpoints_for_method("protected_endpoint",
+                                                          "POST") + self.get_correct_endpoints_for_method(
+            "public_endpoint", "POST")
+        if self.current_protected_endpoint:
+
+            for ep in endpoints:
+                for account in self.accounts:
+                    post_endpoint = ep.get("path")
+                    schema = self.openapi_spec_parser.get_schema_for_endpoint(post_endpoint, "POST")
+                    if "api" in account.keys() and account["api"] in ep:
+                        str_id = f"{account.get('id')}"
+                        post_endpoint = ep.replace("{id}", str_id)
+
+                        prompts = self.test_sql_injection(account, post_endpoint, schema, prompts)
+
+                        prompts = self.test_inputs(post_endpoint, schema, account, prompts)
+
+
+
+                    else:
+                        prompts = self.test_sql_injection(account, post_endpoint, schema, prompts)
+
+                        prompts = self.test_inputs(post_endpoint, schema, account, prompts)
+
+        return prompts
+
+    def generate_xss_prompts(self):
+        """
+           Generate prompts for detecting Cross-Site Scripting (XSS) vulnerabilities.
+
+           This method covers both POST and GET endpoints, targeting public and protected resources.
+           It attempts to inject malicious XSS payloads into input fields (via POST) and query parameters (via GET),
+           and then observes the responses for evidence of unescaped rendering or unsafe HTML reflection.
+
+           Returns:
+               list: A list of prompts designed to detect XSS vulnerabilities.
+           """
+        prompts = []
+        endpoints = self.get_correct_endpoints_for_method("protected_endpoint",
+                                                          "POST") + self.get_correct_endpoints_for_method(
+            "public_endpoint", "POST")
+
+        for account in self.accounts:
+
+            for post_endpoint, _, _ in endpoints:
+                schema = self.openapi_spec_parser.get_schema_for_endpoint(post_endpoint, "POST")
+                prompts = self.test_xss(post_endpoint, account, schema, prompts)
+
+            endpoints = self.get_correct_endpoints_for_method("protected_endpoint",
+                                                              "GET") + self.get_correct_endpoints_for_method(
+                "public_endpoint", "GET")
+            for get_endpoint, _, _ in endpoints:
+                if "id}" in get_endpoint:
+                    get_endpoint = self.replace_placeholders_with_1(get_endpoint, account.get("id"))
+
+                    prompts = self.test_xss_query(get_endpoint, account, prompts)
+
+        return prompts
+
+    def generate_csrf_prompts(self):
+        """
+           Generate prompts to test Cross-Site Request Forgery (CSRF) protection.
+
+           This method tests if sensitive endpoints are protected from unauthorized or forged requests
+           by simulating actions like changing user data without a valid CSRF token.
+           It also checks cookie configurations to ensure proper CSRF defense mechanisms are in place.
+
+           Returns:
+               list: A list of CSRF-related prompts covering POST and GET requests on sensitive endpoints.
+           """
+        prompts = []
+        endpoints = self.get_correct_endpoints_for_method("sensitive_data_endpoint",
+                                                          "POST") + self.get_correct_endpoints_for_method(
+            "sensitive_data_endpoint", "POST")
+        for account in self.accounts:
+            for sensitive_action_endpoint in endpoints:
+                schema = sensitive_action_endpoint.get("schema")
+                prompts = self.test_csrf(sensitive_action_endpoint, schema, prompts, method="POST")
+            endpoints = self.get_correct_endpoints_for_method("sensitive_data_endpoint",
+                                                              "GET") + self.get_correct_endpoints_for_method(
+                "sensitive_data_endpoint", "GET")
+            for sensitive_data_endpoint in endpoints:
+                if "id}" in sensitive_data_endpoint:
+                    sensitive_data_endpoint = self.replace_placeholders_with_1(sensitive_data_endpoint,
+                                                                               account.get("id"))
+                    prompts = self.test_csrf(sensitive_data_endpoint, None, prompts, method="GET")
+
+            # This prompt tests if the API applies CSRF protection to GET requests that handle sensitive data.
+
+            for login in self.login_endpoint:
+
+                login_path, login_schema = self.get_path_and_schema(login)
+                if login_schema is None:
+                    continue
+                if login_path:
+                    prompts = self.test_cookies(login_path, login_schema, prompts, account=account)
+
+        return prompts
+
+    def generate_business_logic_vul_prompts(self):
+        """
+           Generate prompts to test for business logic vulnerabilities.
+
+           These include logic flaws like privilege escalation, incorrect role validation,
+           or bypassing user controls. The method targets both protected and public POST endpoints,
+           as well as sensitive GET endpoints and role-based POST operations.
+
+           Returns:
+               list: A list of prompts to test business rules and role validation logic.
+           """
+        prompts = []
+        endpoints = self.get_correct_endpoints_for_method("protected_endpoint",
+                                                          "POST") + self.get_correct_endpoints_for_method(
+            "public_endpoint", "POST")
+        for account in self.accounts:
+
+            for endpoint, _, _ in endpoints:
+                schema = self.openapi_spec_parser.get_schema_for_endpoint(endpoint, "POST")
+                prompts = self.test_buisness_logic(endpoint, schema, account, prompts, method="POST")
+
+            get_endpoints = self.get_correct_endpoints_for_method("sensitive_data_endpoint", "GET")
+
+            for endpoint, _, _ in get_endpoints:
+                if "id}" in endpoint:
+                    get_endpoint = self.replace_placeholders_with_1(endpoint, account.get("id"))
+                prompts = self.test_buisness_logic(endpoint, None, account, prompts, method="GET")
+            post_endpoints = self.get_correct_endpoints_for_method("role_access_endpoint", "POST")
+
+            for post_endpoint, _, _ in post_endpoints:
+                schema = self.openapi_spec_parser.get_schema_for_endpoint(post_endpoint, "POST")
+                prompts = self.test_roles(post_endpoint, schema, account, prompts)
+
+        return prompts
+
+    def test_injection_bender_account(self, login_path, prompts):
+        injection_account = None
+        bender_account = None
+        if "OWASP" in self.config.get("name"):
+            injection_account = {"email": "' or 1=1--", "password": "a", "x": self.counter}
+            self.counter = self.counter + 1
+            bender_account = self.config.get("bender")
+            bender_account["x"] = self.counter + 1
+        if injection_account is not None:
+            prompts.append({
+                "objective": "SQL Injection Attack on Login (Administrator Account)",
+                "steps": [
+                    f"Send a POST request to {login_path} with user:{injection_account}.\n"
+                ],
+                "token": [""],
+                "path": [login_path],
+                "expected_response_code": [
+                    "400", "404", "403"
+                ],
+                "security": [
+                    "Ensure that input fields like the login form are properly sanitized to prevent SQL injection attacks.",
+                    "Implement prepared statements or parameterized queries to eliminate the possibility of SQL injection exploitation.",
+                    "Verify that authentication systems prevent unauthorized login by enforcing strict checks on input data."
+                ]
+            }
+
+            )
+        if bender_account is not None:
+            prompts.append({
+                "objective": "SQL Injection Attack on Login (Bender Account)",
+                "steps": [
+                    f"Send a POST request to {login_path} with user:{bender_account}.\n"
+                ],
+                "token": [""],
+                "path": [login_path],
+                "expected_response_code": [
+                    "400", "404", "403"
+                ],
+                "security": [
+                    "Ensure that input fields like the login form are properly sanitized to prevent SQL injection attacks.",
+                    "Implement prepared statements or parameterized queries to eliminate the possibility of SQL injection exploitation.",
+                    "Verify that authentication systems prevent unauthorized login by enforcing strict checks on input data."
+                ]
+            }
+            )
+        return prompts
+
+    def extract_change_email_endpoints_with_schema(self) -> list[tuple]:
+        """
+        Extracts all endpoints containing 'change-email' and returns (path, schema) pairs.
+
+        Returns:
+            list: List of tuples (path, schema) for change-email endpoints.
+        """
+        result = []
+
+        for path, methods in self.endpoints.items():
+            if "change-email" in path:
+                for method, method_detail in methods.items():
+                    schema = method_detail.get("requestBody", {}).get("content", {}).get("application/json", {}).get(
+                        "schema")
+                    result.append((path, schema))
+
+        return result
+
+    def test_css(self, endpoint, prompts, schema=None):
+        if schema:
+            prompts.append(
+
+                # Cross-Site Scripting Protection (for OAuth)
+                {
+                    "objective": "Cross-Site Scripting Protection (for OAuth)",
+                    "steps": [
+                        f" Alter Redirect URI in OAuth Request - Send an OAuth authentication request to the endpoint {endpoint} based on schema {schema} with a modified redirect URI that differs from the registered URI. The redirect URI in the request should be altered to a different URL (e.g., https://malicious.site/redirect) to test if the API enforces a strict redirect policy. This step tests if the API strictly validates the redirect URI in OAuth flows, preventing open redirection vulnerabilities.",
+                    ],
+                    "token": [""],
+                    "path": [endpoint],
+                    "expected_response_code": [
+                        "400 Bad Request or 403 Forbidden to indicate successful rejection of unauthorized or mismatched redirect URIs in OAuth requests."
+                    ],
+                    "security": {
+                        [
+                            "Ensure that the OAuth service correctly identifies and rejects requests where the redirect URI does not match the one registered with the client ID. This is crucial to prevent attackers from exploiting open redirection vulnerabilities."
+                            "Verify that the security configuration effectively blocks unauthorized attempts to redirect users to malicious sites through OAuth requests. This protects users from potential cross-site scripting (XSS) attacks and unauthorized data access."]
+                    }
+                }
+
+            )
+        else:
+            prompts.append(
+
+                # Cross-Site Scripting Protection (for OAuth)
+                {
+                    "objective": "Cross-Site Scripting Protection (for OAuth)",
+                    "steps": [
+                        f" Alter Redirect URI in OAuth Request - Send an OAuth authentication request to the endpoint {endpoint} with a modified redirect URI that differs from the registered URI. The redirect URI in the request should be altered to a different URL (e.g., https://malicious.site/redirect) to test if the API enforces a strict redirect policy. This step tests if the API strictly validates the redirect URI in OAuth flows, preventing open redirection vulnerabilities.",
+                    ],
+                    "expected_response_code": [
+                        "400 Bad Request or 403 Forbidden to indicate successful rejection of unauthorized or mismatched redirect URIs in OAuth requests."
+                    ],
+                    "token": [""],
+                    "path": [endpoint],
+                    "security": {
+                        [
+                            "Ensure that the OAuth service correctly identifies and rejects requests where the redirect URI does not match the one registered with the client ID. This is crucial to prevent attackers from exploiting open redirection vulnerabilities."
+                            "Verify that the security configuration effectively blocks unauthorized attempts to redirect users to malicious sites through OAuth requests. This protects users from potential cross-site scripting (XSS) attacks and unauthorized data access."]
+                    }
+                }
+
+            )
+
+        return prompts
+
+    def test_comment(self, account, prompts):
+        comment_endpoint = self.get_comment_ep()
+        if comment_endpoint:
+            prompts = prompts + [
+                {
+                    "objective": "Excessive data exposure",
+                    "steps": [
+                        f"Send a GET request to {comment_endpoint} with a valid token = {account.get('token')}.",
+                    ],
+                    "expected_response_code": [
+                        "200 OK for the initial use of the valid token, confirming it is active.",
+                    ],
+                    "token": [account.get("token")],
+                    "path": [comment_endpoint],
+                    "security":
+                        [
+                            "Ensure that there is only one id in the response."]
+
+                }
+            ]
+        return prompts
+
+    def test_sql_injection(self, account, endpoint, schema, prompts):
+        params = [{"username": "' OR 1=1 #", "password": "password"},
+                  {"username": "'", "password": "password"},
+                  {"username": "admin", "password": "'"}]
+        for param in params:
+            schema = self.fill_schema(schema, param)
+            if schema != {} or len(schema) != 0:
+
+                if "example" in schema:
+                    schema = schema.get("example")
+                    if "{{" in schema:
+                        continue
+                prompts = prompts + [
+                    # Test for Common Injection Vulnerabilities
+                    {
+                        "objective": "Test for Common Injection Vulnerabilities",
+                        "steps": [
+                            f"Send a POST request to exploit common injection vulnerabilities by sending malicious input to {endpoint} based on the schema {schema}."
+                        ],
+                        "token": [account.get("token")],
+                        "path": [endpoint],
+                        "expected_response_code": [
+                            "400 Bad Request,API detects syntactically incorrect inputs due to malformed SQL queries",
+                            "403 Forbidden,API recognizes the input as an unauthorized attempt to access or manipulate data",
+                            "422 Unprocessable Entity (if using WebDAV extensions of HTTP): This is suitable for cases where the request format is correct, but the content includes manipulative commands or data that the server refuses to process."
+                        ],
+                        "security": [
+                            "Ensure the API has measures in place to detect and neutralize injection attacks, securing the application from malicious input manipulation."
+                        ]
+                    }]
+
+        return prompts
+
+    def test_inputs(self, post_endpoint, schema, account, prompts):
+        xss_user = '{"email": "<script>alert(\\"XSS2\\")</script>", "password": ""}'
+        example = {'field1': 'value1', 'field2': 123}
+
+        prompts = prompts + [
+
+            # Test Valid Data Input
+
+            {
+                "objective": "Test Valid Data Input",
+                "steps": [
+                    f"Send a POST request to {post_endpoint} with a payload that matches the valid schema {schema}. This step verifies that the API can correctly process and accept valid data as expected.",
+                ],
+                "expected_response_code": [
+                    "200 OK",
+                    "201 Created indicating that the input data is valid and successfully processed by the API."
+                ],
+                "path": [post_endpoint],
+                "token": [account.get("token")],
+                "security":
+                    [
+                        "Ensure data validation mechanisms are robust to prevent malformed or malicious data entries. This involves confirming that all input adheres to the expected schema and triggers appropriate responses, which is critical for maintaining the integrity and security of the application."
+                    ]}
+            # This request tests that the API processes valid data successfully and as expected.
+            ,
+
+            # Test Invalid Data Input
+            {
+                "objective": "Test Invalid Data Input",
+                "steps": [
+                    f"send a POST request to {post_endpoint} with a payload that violates the schema {schema}. An example of an invalid payload might be: {example}, where data types or required fields are incorrect. This step tests the API's ability to validate data against the schema and handle errors.",
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    [
+                        "400 Bad Request indicating that the API correctly identifies invalid data inputs and rejects them, as per the validation rules defined in the schema."]
+                ],
+                "security":
+                    [
+                        "Ensure that the API's input validation mechanisms are effectively safeguarding against malformed, incorrect, or maliciously crafted data. Robust validation is essential for preventing data integrity issues and securing the API from common vulnerabilities such as injection attacks."
+                    ]}
+
+            # This request tests the API’s response to invalid data, ensuring it properly rejects malformed input.
+            ,
+
+            # Test Edge Case Data Input
+            {
+                "objective": "Test Valid Edge Case Data Input",
+                "steps": [
+                    f"send a POST request to {post_endpoint} with valid edge case values based on the schema {schema}. Examples of valid edge case payloads might include: {{'field1': 'short', 'field2': 1}}, testing the system's handling of minimal valid inputs."
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "200 OK",
+                    "201 Created status, confirming that it can gracefully handle edge cases within the expected boundaries."
+                ],
+                "security":
+                    [
+                        "Ensure that the API's handling of valid edge cases adheres to expected data integrity rules and does not trigger any exceptions or errors, maintaining the reliability and security of the system."
+                    ]},
+            {
+                "objective": "Test Invalid Edge Case Data Input",
+                "steps": [
+                    f"send a POST request to {post_endpoint} with invalid edge case values that clearly violate the boundaries defined by the schema {schema}. Examples of invalid edge case payloads might include: {{'field1': '', 'field2': -999999999999}}, testing the system's ability to reject inputs that are out of acceptable range or format."
+                ],
+                "path": [post_endpoint],
+                "token": [account.get("token")],
+                "expected_response_code": [
+                    "400 Bad Request, confirming that it enforces data constraints and effectively manages inputs that could lead to potential vulnerabilities."
+                ],
+                "security":
+                    [
+                        "Verify that the API's error handling for invalid edge cases prevents data corruption, unhandled exceptions, and security vulnerabilities, such as buffer overflows or other unexpected behaviors that could compromise system stability."
+                    ]}
+
+            # This request tests if the API can handle extreme values or edge cases without errors.
+            ,
+
+            # Test Missing Required Fields
+            {
+                "objective": "Test Missing Required Fields",
+                "steps": [
+                    f"Send a POST request to {post_endpoint} omitting required fields based on {schema}. Example payload: {{'field1': 'value1'}} (missing 'field2'). This tests the API's ability to enforce required fields and provide feedback on missing data."
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "400 Bad Request, specifying the missing fields, confirming that the API enforces data completeness."
+                ],
+                "security": [
+                    "Verify that the API's error handling for missing required fields prevents data processing issues and maintains integrity."
+                ]
+            }
+
+            # This request tests if the API enforces required fields and provides feedback on missing data.
+            ,
+
+            # Test Special Characters and Injection Attacks
+            {
+                "objective": "Test Special Characters and Injection Attacks",
+                "steps": [
+                    f"Send a POST request to {post_endpoint} with potentially malicious data, such as SQL or XSS payloads in fields based on the schema {schema} . Example payload: {{'field1': '<script>alert(1);</script>', 'field2': '1 OR 1=1'}}. This tests the API's ability to sanitize inputs to prevent injection attacks.user: {xss_user}.\n"
+                ],
+                "path": [post_endpoint],
+                "token": [account.get("token")],
+
+                "expected_response_code": [
+                    "200 OK, API sanitizes the input",
+                    "400 Bad Request,API recognizes the input as malicious and rejects the request ",
+                    " 403 Forbidden,API recognizes the input as malicious and rejects the request "
+                ],
+                "security": [
+                    "Check that the API effectively sanitizes and neutralizes special characters and injection vectors to secure data handling."
+                ]
+            }
+
+            # This request tests if the API sanitizes inputs to prevent injection attacks.
+            ,
+
+            # Test Incorrect Data Format
+            {
+                "objective": "Test Special Characters and Injection Attacks",
+                "steps": [
+                    f"Send a POST request to {post_endpoint} with potentially malicious data, such as SQL or XSS payloads in fields based on the schema {schema}. Example payload: {{'field1': '<script>alert(1);</script>', 'field2': '1 OR 1=1'}}. This tests the API's ability to sanitize inputs to prevent injection attacks."
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "200 OK, API sanitied input",
+                    "400 Bad Request, API recognizes input as malicious data.",
+                    "403 Forbidden, API recognizes input as malicious data. "
+                ],
+                "security": [
+                    "Check that the API effectively sanitizes and neutralizes special characters and injection vectors to secure data handling."
+                ]
+            }
+
+            # This request tests if the API validates data formats and rejects incorrect data types.
+            ,
+
+            # General Error Handling and Response Code Check
+            {
+                "objective": "General Error Handling and Response Code Check",
+                "steps": [
+                    f"Send POST requests with a range of invalid inputs to {post_endpoint} to test the API's error handling consistency based on the schema {schema}. Example payloads include missing fields, incorrect types, and special characters. This tests if the API consistently uses appropriate HTTP status codes and provides clear error messages."
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "400 Bad Request for syntactically incorrect data or missing fields",
+                    "422 Unprocessable Entity for correct format but semantically erroneous data (if applicable)"
+                ],
+                "security": [
+                    "Check that the API's error responses do not expose sensitive information while providing helpful error context."
+                ]
+            }
+
+            # This request tests if the API uses consistent HTTP codes and provides clear error messages for invalid inputs.
+            ,
+
+            # Test for Common Injection Vulnerabilities
+            {
+                "objective": "Test for Common Injection Vulnerabilities",
+                "steps": [
+                    f"Send a POST request to exploit common injection vulnerabilities by sending malicious input to {post_endpoint} based on the schema {schema}. Example payload: {{'field1': \"'; DROP TABLE users; --\"}}. This tests the API’s defenses against SQL injection and other forms of injection."
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "400 Bad Request,API detects syntactically incorrect inputs due to malformed SQL queries",
+                    "403 Forbidden,API recognizes the input as an unauthorized attempt to access or manipulate data",
+                    "422 Unprocessable Entity (if using WebDAV extensions of HTTP): This is suitable for cases where the request format is correct, but the content includes manipulative commands or data that the server refuses to process."
+                ],
+                "security": [
+                    "Ensure the API has measures in place to detect and neutralize injection attacks, securing the application from malicious input manipulation."
+                ]
+            }
+
+            # This request tests the API’s resistance to injection attacks, ensuring input handling is secure.
+            ,
+
+            # Validate Data Sanitization and Security
+            {
+                "objective": "Validate Data Sanitization and Security",
+                "steps": [
+                    f"Send a POST request with data containing special characters and potentially harmful inputs to {post_endpoint} based on the schema {schema}. Confirm that the API sanitizes and escapes all data properly."
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "Ensure no vulnerabilities are exposed through unsanitized inputs."
+                ],
+                "security": [
+                    "Verify comprehensive sanitization across all input fields to prevent security risks associated with unescaped data."
+                ]
+            }
+
+            # This request checks that the API applies appropriate sanitization to user-supplied data across all fields.
+
+        ]
+
+        return prompts
+
+    def test_xss(self, post_endpoint, account, schema, prompts):
+        prompts = prompts + [
+            # Test for XSS by Injecting Script into Text Fields
+            {
+                "objective": "Test for XSS by Injecting Script into Text Fields",
+                "steps": [
+                    f"Send a POST request to {post_endpoint} where user input is accepted (e.g., comment or message fields) based on schema {schema}."
+                    "Include a payload with a script tag, such as <script>alert('XSS');</script>, in the request body to test if the API escapes or sanitizes it correctly."
+                ],
+                "path": [post_endpoint],
+                "token": [account.get("token")],
+
+                "expected_response_code": [
+                    "200 OK if the input is safely sanitized and stored.",
+                    "400 Bad Request if the payload is outright rejected due to validation rules."
+                ],
+                "security": [
+                    "Verify that the API escapes or sanitizes all user inputs to prevent reflected XSS attacks.",
+                    "Ensure that the response does not include any unescaped or executable scripts."
+                ]
+            }
+
+            # This prompt tests if the API properly sanitizes user inputs in text fields to prevent reflected XSS attacks.
+            ,  # Test for Persistent XSS
+            {
+                "objective": "Test for Persistent XSS",
+                "steps": [
+                    f"Send a POST request to {post_endpoint} to store user-provided data (e.g., profile descriptions or comments) based on schema {schema}."
+                    "Inject a persistent XSS payload, like <script>alert('Persistent XSS');</script>, in the request body."
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "200 OK if the input is safely sanitized and stored without executing any scripts.",
+                    "400 Bad Request if the input is rejected due to validation rules."
+                ],
+                "security": [
+                    "Ensure that the API sanitizes or escapes all stored user inputs to prevent stored XSS attacks.",
+                    "Verify that scripts in stored content are not executed when retrieved."
+                ]
+            }
+
+            # This prompt tests if the API prevents stored XSS by sanitizing inputs that might be displayed to other users.
+            ,
+
+        ]
+        return prompts
+
+    def test_xss_query(self, get_endpoint, account, prompts):
+        prompts = prompts + [
+            # Check for XSS in Query Parameters
+            {
+                "objective": "Check for XSS in Query Parameters",
+                "steps": [
+                    f"Send a GET request to {get_endpoint} with a query parameter containing a script payload, such as ?search=<script>alert('XSS');</script>.",
+                ],
+                "token": [account.get("token")],
+                "path": [get_endpoint],
+                "expected_response_code": [
+                    "200 OK if the payload is properly sanitized and reflected without execution.",
+                    "400 Bad Request if the API blocks the input due to validation rules."
+                ],
+                "security": [
+                    "Ensure that all query parameters are sanitized or escaped to prevent reflected XSS attacks.",
+                    "Verify that the response does not contain unescaped scripts or unsafe data."
+                ]
+            }
+
+            # This prompt tests if the API properly handles and escapes query parameters to prevent reflected XSS.
+            ,
+            # Test for XSS in JSON Responses
+            {
+                "objective": "Test for XSS in JSON Responses",
+                "steps": [
+                    f"Send a GET request to {get_endpoint} expecting a JSON response."
+                    "Inject a payload in a query parameter (e.g., ?param=<script>alert('XSS');</script>) and observe if the JSON response reflects this payload without escaping.",
+                ],
+                "token": [account.get("token")],
+                "path": [get_endpoint],
+                "expected_response_code": [
+                    "200 OK if the JSON response properly escapes or removes the malicious payload.",
+                    "400 Bad Request if the input is rejected due to validation rules."
+                ],
+                "security": [
+                    "Verify that JSON responses properly escape all user inputs to prevent XSS vulnerabilities.",
+                    "Ensure that malicious scripts in parameters are not reflected in the response."
+                ]
+            }
+
+            # This prompt tests if JSON responses are properly escaped, ensuring no XSS vulnerabilities when displaying data to users.
+            ,
+            # Inspect Headers and Metadata for XSS Protections
+            {
+                "objective": "Inspect Headers and Metadata for XSS Protections",
+                "steps": [
+                    f"Send a GET request to {get_endpoint}.",
+                ],
+                "path": [get_endpoint],
+                "token": [account.get("token")],
+                "expected_response_code": [
+                    "200 OK with headers such as Content-Security-Policy and X-XSS-Protection properly set."
+                ],
+                "security": [
+                    "Inspect the response headers for security settings such as Content-Security-Policy, X-XSS-Protection, and others to verify if XSS protections are enabled."
+                    "Ensure that the API response headers contain adequate security settings to mitigate XSS risks.",
+                    "Verify that the Content-Security-Policy header is configured to prevent inline scripts or unauthorized domains from executing scripts."
+                ]
+            }
+
+            # This prompt checks if appropriate security headers are set to protect against XSS vulnerabilities.
+        ]
+
+        return prompts
+
+    def test_csrf(self, endpoint, schema, prompts, method="POST"):
+        if method == "POST":
+            prompts = prompts + [
+                # Test for CSRF Protection in Sensitive Actions
+                {
+                    "objective": "Test for CSRF Protection in Sensitive Actions",
+                    "steps": [
+                        f"Send a POST request to {endpoint}  based on schema {schema}(e.g., account update or password change) without including any authentication tokens (e.g., CSRF token or session cookie).",
+                    ],
+                    "token": [""],
+                    "path": [endpoint],
+                    "expected_response_code": [
+                        "403 Forbidden if CSRF protection is correctly enforced and the request lacks proper tokens.",
+                        "401 Unauthorized if authentication is required and missing."
+                    ],
+                    "security": [
+                        "Observe the API's response to determine if the request is rejected due to missing tokens."
+                        "Ensure that the API requires authentication tokens for sensitive actions to mitigate CSRF attacks.",
+                        "Verify that the absence of a CSRF token results in a proper rejection of the request."
+                    ]
+                }
+
+                # This prompt tests if the API protects sensitive actions by requiring proper authentication tokens.
+                ,  # Test for CSRF Token Requirement
+                {
+                    "objective": "Test for CSRF Token Requirement",
+                    "steps": [
+                        f"Send a POST request to {endpoint} based on schema {schema} with valid session cookies but without including a CSRF token.",
+                    ],
+                    "token": [""],
+                    "path": [endpoint],
+                    "expected_response_code": [
+                        "403 Forbidden if the API correctly requires a CSRF token for the action.",
+                        "400 Bad Request if the request format is invalid or missing required headers."
+                    ],
+                    "security": [
+                        "Check if the API rejects the request, indicating that CSRF tokens are mandatory for protected actions."
+                        "Confirm that CSRF tokens are required for all sensitive actions and that their absence results in the rejection of requests.",
+                        "Ensure that proper error messages are returned without exposing sensitive system details."
+                    ]
+                }
+
+                # This prompt verifies if the API requires a CSRF token, ensuring protection against CSRF attacks.
+                ,  # Test for CSRF Token Validation
+                {
+                    "objective": "Test for CSRF Token Validation",
+                    "steps": [
+                        f"Send a POST request to {endpoint} based on schema {schema} with valid session cookies but an incorrect or expired CSRF token.",
+                    ],
+                    "path": [endpoint],
+                    "token": [self.create_random_bearer_token()],
+                    "expected_response_code": ["403 Forbidden if the CSRF token is invalid or expired.",
+                                               "401 Unauthorized if the session cookies are also invalid or missing."],
+                    "security": [
+                        "Verify if the API rejects the request, indicating that the provided CSRF token is invalid.",
+                        "Ensure that the API validates the CSRF token correctly, rejecting requests with invalid or expired tokens.",
+                        "Verify that error messages do not reveal sensitive information about the token validation process."]}
+
+                # This prompt tests if the API validates the CSRF token, ensuring only valid tokens are accepted.
+                , ]
+        elif method == "GET":
+            prompts = prompts + [
+                # Test CSRF Protection in GET Requests
+
+                {
+                    "objective": "Test CSRF Protection in GET Requests",
+                    "steps": [
+                        f"Send a GET request to {endpoint} to retrieve sensitive data (e.g., personal information) without including a CSRF token or session cookie.",
+                    ],
+                    "token": [""],
+                    "path": [endpoint],
+                    "expected_response_code": [
+                        "401 Unauthorized if authentication is required but missing.",
+                        "403 Forbidden if the API enforces CSRF protection for sensitive GET requests."
+                    ],
+                    "security": [
+                        "Observe the response to check if the API enforces CSRF protection or authentication for GET requests with sensitive data.",
+                        "Confirm that the API applies CSRF protection or requires authentication for GET requests that handle sensitive data.",
+                        "Ensure that unauthorized access attempts do not result in any sensitive data exposure."
+                    ]
+                }
+            ]
+
+            return prompts
+
+    def test_buisness_logic(self, endpoint, schema, account, prompts, method="POST"):
+        if method == "POST":
+            prompts.append(
+                # Test for Unauthorized Access Through Business Logic
+                {
+                    "objective": "Test for Unauthorized Access Through Business Logic",
+                    "steps": [
+                        f"Send a POST request to {endpoint} based on schema {schema} with manipulated data, such as setting an unauthorized discount or modifying the item price with user:{account}.\n",
+                    ],
+                    "path": [endpoint],
+                    "token": [account.get("token")],
+                    "expected_response_code": [
+                        "403 Forbidden if the API correctly identifies and prevents unauthorized actions.",
+                        "400 Bad Request if the input data violates business rules or constraints."
+                    ],
+                    "security": [
+                        "Verify if the API prevents the manipulated changes from being applied and restricts unauthorized actions.",
+                        "Ensure the API enforces proper validation and checks on critical business logic operations to prevent unauthorized manipulations.",
+                        "Verify that manipulated requests do not bypass security measures or result in unauthorized actions."
+                    ]
+                }
+
+                # This prompt tests for flaws in the business logic that could allow unauthorized discounts or price manipulations.
+            )
+        elif method == "GET":
+            prompts.append(
+                # Test for Data Exposure via Business Logic Flaws
+                {
+                    "objective": "Test for Data Exposure via Business Logic Flaws",
+                    "steps": [
+                        f"Send a GET request to {endpoint} as a low-privilege user:{account}.\n",
+
+                    ],
+                    "path": [endpoint],
+                    "token": [account.get("token")],
+                    "expected_response_code": [
+                        "403 Forbidden if the user lacks permission to access the requested data.",
+                        "200 OK only if the user is authorized to access the data."
+                    ],
+                    "security": [
+                        "Attempt to retrieve sensitive information that should not be accessible to the user.",
+                        "Verify if the API enforces permission checks to restrict access to data beyond the user's allowed level."
+                        "Ensure the API applies strict access controls to sensitive data based on user permissions.",
+                        "Verify that unauthorized access attempts are logged and handled securely."
+                    ]
+                }
+
+                # This prompt tests if the business logic prevents unauthorized access to sensitive user data.
+            )
+
+        return prompts
+
