+++ src/hackingBuddyGPT/utils/prompt_generation/information/_pentesting_misc.py
@@ -1,0 +1,359 @@
+# This file is part of the PenTestingInformation split (see pentesting_information.py).
+import base64
+import copy
+import csv
+import glob
+import os
+import random
+import re
+import secrets
+from typing import List
+
+from hackingBuddyGPT.utils.openapi.openapi_parser import OpenAPISpecificationParser
+from hackingBuddyGPT.utils.prompt_generation.information.prompt_information import (
+    PromptPurpose,
+)
+from faker import Faker
+
+
+class MiscScenarios:
+    """MiscScenarios: scenario/helper methods mixed into PenTestingInformation."""
+
+    def generate_error_handling_prompts(self):
+        """
+          Generate prompts for testing error handling on POST endpoints.
+
+          This method verifies that endpoints respond with meaningful and secure error messages
+          when provided with incorrect or malformed input.
+
+          It combines protected and public POST endpoints, retrieves their schemas, and uses
+          account information to inject malformed or edge-case data to observe error behavior.
+
+          Returns:
+              list: A list of prompts to test the robustness and clarity of error handling.
+          """
+        prompts = []
+        endpoints = self.get_correct_endpoints_for_method("protected_endpoint",
+                                                          "POST") + self.get_correct_endpoints_for_method(
+            "public_endpoint", "POST")
+
+        for ep in endpoints:
+            post_endpoint = ep.get("path")
+            schema = self.openapi_spec_parser.get_schema_for_endpoint(post_endpoint, "POST")
+            for account in self.accounts:
+                if "api" in post_endpoint and "api" in account.keys() and account["api"] in ep:
+                    str_id = f"{account.get('id')}"
+                    post_endpoint = ep.replace("{id}", str_id)
+
+                    prompts = self.test_error_handling(post_endpoint, account, schema, prompts)
+
+
+                else:
+                    prompts = self.test_error_handling(post_endpoint, account, schema, prompts)
+
+        return prompts
+
+    def generate_security_misconfiguration_prompts(self):
+        """
+          Generate prompts for identifying security misconfigurations in the API.
+
+          This method checks for:
+          - Missing or misconfigured security-related HTTP headers (e.g., CSP, X-Frame-Options).
+          - Insecure key exposure through endpoint patterns.
+          - Incorrect CORS settings that may lead to data leakage.
+          - Disclosure of sensitive debug or configuration information in error responses.
+
+          Returns:
+              list: A list of prompts targeting potential security misconfiguration vulnerabilities.
+          """
+        prompts = []
+
+        user_endpoints = [ep for ep in self.endpoints if ep.endswith("user")]
+
+        for key_endpoint in user_endpoints:
+            for account in self.accounts:
+                if account["api"] in key_endpoint:
+                    if "key" in account:
+                        prompts = self.test_limiting("", key_endpoint, prompts, account.get("key"))
+
+        endpoints = self.get_correct_endpoints_for_method("protected_endpoint",
+                                                          "GET") + self.get_correct_endpoints_for_method(
+            "public_endpoint", "GET")
+        user_endpoints = [ep for ep in self.endpoints if "key" in ep]
+        endpoints = endpoints + user_endpoints
+
+        for endpoint in endpoints:
+            if isinstance(endpoint, dict):
+                endpoint = endpoint.get("path")
+            for account in self.accounts:
+                id = account.get("id")
+                if f"{id}".isdigit() and "{id}" in endpoint:
+                    endpoint = endpoint.replace("{id}", f"{id}")
+                prompts.append(
+
+                    # Check for Security Misconfigurations in Headers
+                    {
+                        "objective": "Check for Security Misconfigurations in Headers",
+                        "steps": [
+                            f"Send a GET request to {endpoint}.",
+
+                        ],
+                        "path": [endpoint],
+                        "token": [account.get("token")],
+                        "expected_response_code": [
+                            "200 OK with response headers correctly set to include security-related configurations such as X-Content-Type-Options: nosniff, X-Frame-Options: SAMEORIGIN, and Content-Security-Policy."
+                        ],
+                        "security": [
+                            "Inspect the response headers, including X-Content-Type-Options, X-Frame-Options, and Content-Security-Policy.",
+                            "Verify if these headers are properly configured to mitigate security risks."
+                            "Ensure that response headers are configured to prevent security vulnerabilities such as clickjacking, MIME type sniffing, and XSS.",
+                            "Verify that no unnecessary or insecure headers are present in the response."
+                            "Check if 'Access-Control-Allow-Origin: * Access-Control-Allow-Credentials: true' is set as these headers indicate that the CORS policy for the given request is misconfigured, allowing it to be invoked from a third-party web application to access the resource and obtain sensitive information like the user's ID, username, password, and auth key."
+                        ]
+                    }
+
+                    # This prompt tests if security headers are properly configured to prevent exposure to security risks.
+                )
+        endpoints = self.get_correct_endpoints_for_method("secure_action_endpoint",
+                                                          "POST")
+        for endpoint in endpoints:
+            for account in self.accounts:
+                id = account.get("id")
+                endpoint = endpoint.replace("{id}", f"{id}")
+
+                schema = self.openapi_spec_parser.get_schema_for_endpoint(endpoint, "POST")
+                for account in self.accounts:
+                    prompts.append(
+                        # Inspect API for Debug or Error Information Exposure
+                        {
+                            "objective": "Inspect API for Debug or Error Information Exposure",
+                            "steps": [
+                                f"Send an invalid POST request to {endpoint} based on schema {schema}.",
+
+                            ],
+                            "token": [account.get("token")],
+                            "path": [endpoint],
+                            "expected_response_code": [
+                                "400 Bad Request or 403 Forbidden with a generic error message that does not disclose sensitive internal details."
+                            ],
+                            "security": [
+                                "Check if the response contains detailed error messages or debug information.",
+                                "Verify that the response provides only minimal and generic error messages."
+                                "Ensure that the API avoids exposing debug or configuration details in error responses.",
+                                "Verify that sensitive system information is not included in error messages."
+                            ]
+                        }
+
+                        # This prompt tests if the API avoids exposing sensitive configuration or debug information.
+                    )
+
+        return prompts
+
+    def generate_logging_monitoring_prompts(self):
+        """
+          Generate prompts to test whether the API performs proper logging and monitoring.
+
+          This method simulates:
+          - Invalid request submissions to see if they are logged appropriately.
+          - Potentially malicious requests (e.g., SQL injection attempts).
+          - Logging of unexpected or unauthorized access attempts.
+
+          It ensures that logs are meaningful, timely, and privacy-compliant without exposing sensitive data.
+
+          Returns:
+              list: A list of prompts to verify proper logging and monitoring mechanisms.
+          """
+        prompts = []
+        endpoints = self.get_correct_endpoints_for_method("secure_action_endpoint",
+                                                          "POST")
+        flag_endpoints = [ep for ep in self.endpoints if "flag" in ep]
+        endpoints = flag_endpoints + endpoints
+
+        for endpoint in endpoints:
+            if isinstance(endpoint, dict):
+                endpoint = endpoint.get("path")
+            schema = self.openapi_spec_parser.get_schema_for_endpoint(endpoint, "POST")
+            for account in self.accounts:
+                prompts.append(
+                    # Test Logging for Incorrect Requests
+                    {
+                        "objective": "Test Logging for Incorrect Requests",
+                        "steps": [
+                            f"Send an invalid POST request to {endpoint} based on schema {schema}.",
+                        ],
+                        "path": [endpoint],
+                        "token": [account.get("token")],
+                        "expected_response_code": [
+                            "400 Bad Request or 403 Forbidden, with the API logging details of the failed attempt."
+                        ],
+                        "security": [
+                            "Verify if the API logs the details of the failed request, including timestamp, user information, and the nature of the error."
+                            "Ensure that the API logs invalid requests comprehensively to assist in debugging and identifying potential misuse.",
+                            "Verify that logs do not include sensitive user data or details beyond what is necessary for monitoring."
+                        ]
+                    }
+
+                    # This prompt tests if the API logs incorrect requests, including any errors encountered.
+                )
+        endpoints = self.get_correct_endpoints_for_method("secure_data_endpoint",
+                                                          "GET")
+        params = [{"username": "' OR 1=1 #", "password": "password"}, {"username": "'", "password": "password"},
+                  {"username": "admin", "password": "'"}]
+
+        for endpoint in endpoints:
+            if isinstance(endpoint, dict):
+                endpoint = endpoint.get("path")
+
+            for account in self.accounts:
+                if "id}" in endpoint:
+                    endpoint = self.replace_placeholders_with_1(endpoint, account.get("id"))
+                    for paramerts in params:
+                        prompts.append(
+                            # Test Logging for Potentially Malicious Requests
+                            {
+                                "objective": "Test Logging for Potentially Malicious Requests",
+                                "steps": [
+                                    f"Send a GET request to {endpoint} with unusual or malicious parameters, such as SQL injection attempts. Request body should look like this: {paramerts}",
+                                ],
+                                "token": [account.get("token")],
+                                "path": [endpoint],
+                                "expected_response_code": [
+                                    "200 OK",
+                                    "400 Bad Request with the API logging the details of the suspicious request, including the malicious parameters and IP address."
+                                ],
+                                "security": [
+                                    "Check if the API logs the suspicious request and flags it for further analysis or monitoring."
+                                    "Ensure that the API logs potentially malicious requests and flags them for review to mitigate security threats.",
+                                    "Verify that logs are monitored for patterns of abuse or repeated attacks."
+                                ]
+                            }
+
+                            # This prompt tests if the API logs and monitors potentially malicious requests to help detect and respond to security threats.
+                        )
+        return prompts
+
+    def test_error_handling(self, post_endpoint, account, schema, prompts):
+        prompts = prompts + [
+
+            # Test Error Handling with Invalid Input
+            {
+                "objective": "Test Error Handling with Invalid Input",
+                "steps": [
+                    f"Send a POST request to {post_endpoint} with deliberately invalid input to test error handling capabilities based on the schema {schema}. Example payload: {{'nonexistent_field': 'invalid_value'}}. This tests the API's ability to recognize and reject invalid input formats or unexpected fields."
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "400 Bad Request, confirming that the API correctly identifies invalid input and provides a clear, informative error message indicating why the input is invalid."
+                ],
+                "security": [
+                    "Ensure the API's error handling mechanisms do not expose sensitive information while adequately informing the user about the nature of the error. This is crucial for maintaining user trust and preventing potential security issues related to erroneous input handling."
+                ]
+            }
+
+            # This request tests if the API correctly handles invalid input with appropriate error messages and status codes.
+            ,
+
+            # Test Error Handling with Missing Required Fields
+            {
+                "objective": "Test Error Handling for Missing Required Fields",
+                "steps": [
+                    f"Send a POST request to {post_endpoint} omitting one or more required fields to test the API’s validation mechanisms based on the schema {schema}. Example payload: {{'optional_field': 'test_value'}}, where only optional fields are provided, and required fields are deliberately left out."
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "400 Bad Request, confirming that the API correctly identifies the lack of required fields and responds with a detailed error message that lists the missing required fields."
+                ],
+                "security": [
+                    "Ensure that the API's error responses for missing required fields are specific enough to guide the user but do not disclose sensitive system information that could be exploited."
+                ]
+            }
+
+            # This request checks if the API enforces required fields and gives helpful feedback on missing data.
+            ,
+
+            # Examine Error Message Detail for Information Leakage
+            {
+                "objective": "Examine Error Message Detail for Information Leakage",
+                "steps": [
+                    f"Send a POST request to {post_endpoint} with incorrect field types or values to provoke an error response from the API v. Example payload: {{'field': 123}}, where the API is expected to receive a string instead. This step tests the API's handling of data type errors and its error messaging security."
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "400 Bad Request, and scrutinize the error message and headers to ensure they do not disclose sensitive information such as stack traces, server configurations, or detailed server internals."
+                ],
+                "security": [
+                    "Assess the security posture of the API by checking that error responses are safely managed. Error messages and headers should convey enough information for user correction without revealing details that could be used maliciously. Confirm that the error responses maintain privacy and security by not exposing detailed system data."
+                ]
+            }
+
+            # This request tests if the API avoids leaking sensitive information through error messages or headers.
+            ,
+
+            # Test Error Response for Security Vulnerabilities
+            {
+                "objective": "Test Error Response for Security Vulnerabilities",
+                "steps": [
+                    f"Send various malformed requests to {post_endpoint} to assess the security of the API's error handling based on the schema {schema}. Example of a malformed request: {{field': 'value' (note the missing closing brace). This tests if the API avoids exposing internal structures or sensitive information in its responses to such requests.}}"
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "400 Bad Request, without any specific error details that could assist in crafting further attacks."
+                ],
+                "security": [
+                    "Ensure that the API responds with generic error messages that do not disclose any details about the internal workings or structures of the system. Appropriate responses might include a , "
+                    "Confirm that error messages are standardized and vague enough to prevent revealing any technical details of the backend systems, such as software versions, frameworks, or detailed error diagnostics. This helps in mitigating risks associated with information leakage."
+                ]
+            }
+
+            # This request confirms that malformed requests are handled securely and don’t expose implementation details.
+            ,
+
+            # Assess Error Messages for Corrective Action Suggestions
+            {
+                "objective": "Assess Error Messages for Corrective Action Suggestions",
+                "steps": [
+                    f"Send a POST request to {post_endpoint} with intentionally incorrect inputs such as missing required fields or wrong data types based on the schema {schema}. Example payload could include: {{'field1': 'value1'}} with 'field2' missing or incorrectly formatted. This tests the API's ability to not only catch errors but also guide users towards resolving them."
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "400 Bad Request, fails validation checks such as missing required fields or incorrect data types",
+                    "422 Unprocessable Entity, entity is correct but was unable to process the contained instruction"
+                ],
+                "security": [
+                    "Verify that the API's error messages are informative and instructive, providing specific suggestions for correcting the input. Expect responses such as 400 Bad Request that include details on what is missing or incorrect, like specifying missing fields or incorrect data types."
+                    "Ensure that while the error messages are helpful, they do not disclose sensitive information or details about the underlying implementation that could be exploited by an attacker. The messages should strike a balance between being helpful and maintaining security."
+                ]
+            }
+
+            # This request tests if the API offers actionable feedback to help users correct their requests.
+            ,
+
+            # Look for Information Leakage Vulnerabilities
+            {
+                "objective": "Look for Information Leakage Vulnerabilities",
+                "steps": [
+                    f"Send POST requests to {post_endpoint} with unexpected input, such as special characters or SQL-like strings, to test how securely the API handles error-prone or potentially malicious data based on the schema {schema}. Example payloads could include: {{'field': \"' OR '1'='1\"}} or {{'field': '<script>alert(\"test\");</script>'}}. This tests the API's resilience against attempts to expose sensitive information."
+                ],
+                "token": [account.get("token")],
+                "path": [post_endpoint],
+                "expected_response_code": [
+                    "400 Bad Request, indicating that the input was syntactically incorrect or otherwise unacceptable,",
+                    "500 Internal Server Error, if unexpected input leads to unhandled errors within the server"
+                ],
+                "security": [
+                    "Inspect if the response contains any server details, stack traces, or database error messages that could be exploited. The API should not reveal such information under any circumstance. Responses should be sanitized to prevent leakage of sensitive details about the server or the application's internal mechanics.",
+                    "Ensure that the error handling is secure and does not inadvertently disclose sensitive information through error messages or HTTP headers. Error responses should be generic and non-descriptive to avoid aiding potential attackers."
+                ]
+            }
+
+            # This request verifies that the API is secure against accidental exposure of sensitive information in responses.
+
+        ]
+
+        return prompts
+
