+++ src/hackingBuddyGPT/utils/prompt_generation/information/_pentesting_ratelimit.py
@@ -1,0 +1,195 @@
+# This file is part of the PenTestingInformation split (see pentesting_information.py).
+import base64
+import copy
+import csv
+import glob
+import os
+import random
+import re
+import secrets
+from typing import List
+
+from hackingBuddyGPT.utils.openapi.openapi_parser import OpenAPISpecificationParser
+from hackingBuddyGPT.utils.prompt_generation.information.prompt_information import (
+    PromptPurpose,
+)
+from faker import Faker
+
+
+class RateLimitScenarios:
+    """RateLimitScenarios: scenario/helper methods mixed into PenTestingInformation."""
+
+    def generate_rate_limit_throttling(self):
+        """
+        Generate prompts to test rate limiting and throttling protections.
+
+        This method simulates high-frequency requests to login and OTP endpoints
+        to detect absence of proper rate-limiting mechanisms, which could lead to
+        brute-force attacks or denial-of-service (DoS).
+
+        Returns:
+            list: A list of prompts that test for request throttling, OTP abuse, and rate limits.
+        """
+        prompts = []
+        for login in self.login_endpoint:
+
+            login_path, login_schema = self.get_path_and_schema(login)
+            request = [
+                          f"Send a large number of POST requests to {login_path} based on schema {login_schema} within a short timeframe."] * 10
+            if login_path:
+                prompts = self.test_limiting(request, login_path, prompts)
+
+        # OPT test cases
+        opt_endpoints = [ep for ep in self.endpoints if "otp" in ep]
+
+        for opt_endpoint in opt_endpoints:
+            numbers = list(range(0000, 10000))
+
+            # Pick 10 random numbers from the list
+            random_numbers = random.sample(numbers, 10)
+            random_numbers = [1872] + random_numbers
+            for number in random_numbers:
+                otp = {"otp": f"{number}"}
+
+                prompts = prompts + [
+
+                    {
+                        "objective": "Test for Unrestricted Resource Consumption",
+                        "steps": [
+                            f"Send a  POST request to trigger the sending of an OTP to a mobile phone number. This is intended to test the system's resilience against DoS attacks through resource exhaustion. Request should look like the following:{otp}"
+                        ],
+                        "token": [""],
+                        # Assuming no authentication is required for this endpoint, or specify where needed
+                        "path": [opt_endpoint],  # Specify the actual endpoint for sending OTP
+                        "expected_response_code": [
+                            "403 Forbidden"
+                        ],
+                        "security": [
+                            "Verify that msg is not '4 Digit OPT sent on mobile no.'",
+
+                        ]
+                    }
+                    ,
+                    {
+                        "objective": "Testing for Unrestricted Resource Consumption",
+                        "steps": [
+                            f"Send a  POST requests to trigger the verify an OTP to a mobile phone number. Request should look like the following:{otp}"
+                        ],
+                        "token": [""],
+                        # Assuming no authentication is required for this endpoint, or specify where needed
+                        "path": [opt_endpoint],  # Specify the actual endpoint for sending OTP
+                        "expected_response_code": [
+                            "403 Forbidden"
+                        ],
+                        "security": [
+                            "Verify that request is okay",
+
+                        ]
+                    }
+                ]
+
+        prompts = self.test_throttling(prompts)
+        return prompts
+
+    def test_limiting(self, request, endpoint, prompts, key=None):
+        if key:
+
+            prompts.append(
+
+                # Check for Lack of Resources & Rate Limiting
+                {
+                    "objective": "Check for Lack of Resources & Rate Limiting",
+                    "steps": [
+                        f"Send a GET request to {endpoint}.",
+
+                    ],
+                    "path": [endpoint],
+                    "token": [key],
+                    "expected_response_code": [
+                        "429 Too many requests"
+                    ],
+                    "security": [
+                        "Ensure each step is followed without skipping to maintain the integrity of the test.",
+                        "Check for HTTP 429 response to confirm if rate limiting is enforced after numerous sequential OTP verification attempts.",
+                        "Assess server response under load for stability and consistency to prevent denial of service.",
+                        "Review the application logs to ensure that all brute force attempts are recorded for security monitoring.",
+                        "Implement rate limiting on sensitive actions like OTP verification to mitigate brute-force attacks."
+                    ]
+                }
+
+                # This prompt tests if security headers are properly configured to prevent exposure to security risks.
+            )
+        else:
+            prompts = prompts + [
+                # Test Rate-Limiting for Abuse Prevention
+                {
+                    "objective": "Test Rate-Limiting for Abuse Prevention",
+                    "steps":
+                        request
+
+                    ,
+                    "expected_response_code": [
+                        "429 Too Many Requests if the API correctly enforces rate-limiting.",
+                        "200 OK for requests that fall within the allowed rate limit."
+                    ],
+                    "path": [endpoint] * 10,
+                    "token": [""] * 10,
+                    "security": [
+                        "Monitor the responses to verify if the API enforces a rate limit after a specific number of requests.",
+                        "Check if the API responds with a 429 Too Many Requests status when the rate limit is exceeded."
+                        "Ensure that the rate-limiting mechanism effectively prevents abuse while allowing legitimate users to access the API without disruption.",
+                        "Verify that rate-limiting thresholds and error messages do not expose sensitive implementation details."
+                    ]
+                }
+
+                # This prompt tests if the API has adequate rate-limiting to prevent abuse.
+            ]
+        return prompts
+
+    def test_throttling(self, prompts):
+        resource_intensive_endpoints = self.get_correct_endpoints_for_method("resource_intensive_endpoint", "GET")
+        for resource_intensive_endpoint, _, _ in resource_intensive_endpoints:
+            if "id}" in resource_intensive_endpoint:
+                resource_intensive_endpoint = self.replace_placeholders_with_1(resource_intensive_endpoint, 1)
+            one = [
+                      f"Send a series of GET requests per second to the endpoint {resource_intensive_endpoint} for 30 seconds and verify that all requests return a 200 OK response."] * 10
+            hundert = [
+                          f"Send a burst of 100 GET requests per second to the endpoint {resource_intensive_endpoint} for 30 seconds and observe the API's response."] * 20
+            ten_status = ["200 OK for all legitimate requests under normal traffic conditions.", ] * 10
+            hundert_status = [
+                                 "429 Too Many Requests if throttling is applied when excessive requests are detected."] * 20
+
+            paths = [resource_intensive_endpoint] * 30
+            tokens = [""] * 30
+
+            prompts.append(
+
+                # Test Throttling for Denial-of-Service (DoS) Attack Prevention
+
+                {
+                    "objective": "Test Throttling for Denial-of-Service (DoS) Attack Prevention",
+                    "steps": [
+                        one,
+                        hundert,
+                        f"Analyze the API's response during the excessive traffic burst. Check for status codes such as 429 Too Many Requests or 503 Service Unavailable, indicating that throttling is applied.",
+                        f"4: Reduce the request rate back to 10 GET requests per second to {resource_intensive_endpoint} and verify that legitimate traffic returns a 200 OK response, confirming the system has recovered."
+                    ],
+                    "expected_response_code": [ten_status,
+                                               hundert_status,
+                                               "503 Service Unavailable if the server temporarily denies access due to heavy load.",
+                                               "4: 200 OK for legitimate requests after the request rate returns to normal."
+                                               ],
+                    "path": paths,
+                    "token": tokens,
+                    "security": [
+                        "Ensure that the API's throttling mechanism effectively prevents DoS attacks by limiting excessive traffic.",
+                        "Verify that throttling does not unintentionally block or degrade service for legitimate users.",
+                        "Confirm that the API recovers quickly and reliably after excessive traffic subsides, maintaining availability for normal traffic."
+                    ]
+                }
+
+                # This prompt tests if the API prevents DoS attacks through request throttling.
+            )
+
+            return prompts
+
