+++ src/hackingBuddyGPT/utils/prompt_generation/information/_pentesting_session.py
@@ -1,0 +1,234 @@
+# This file is part of the PenTestingInformation split (see pentesting_information.py).
+import base64
+import copy
+import csv
+import glob
+import os
+import random
+import re
+import secrets
+from typing import List
+
+from hackingBuddyGPT.utils.openapi.openapi_parser import OpenAPISpecificationParser
+from hackingBuddyGPT.utils.prompt_generation.information.prompt_information import (
+    PromptPurpose,
+)
+from faker import Faker
+
+
+class SessionScenarios:
+    """SessionScenarios: scenario/helper methods mixed into PenTestingInformation."""
+
+    def generate_session_management_prompts(self):
+        """
+           Generate prompts for testing session management and security.
+
+           This method checks GET endpoints (both protected and public) for:
+           - Proper session validation.
+           - Session hijacking resistance.
+           - Session-related cookie attributes (e.g., HttpOnly, Secure).
+           - Other session vulnerabilities.
+
+           It also evaluates login endpoints to simulate authentication flows and test
+           how sessions are managed, maintained, and secured afterward.
+
+           Returns:
+               list: A list of prompts testing session integrity, hijacking protections, and cookie configurations.
+           """
+        prompts = []
+
+        endpoints = self.get_correct_endpoints_for_method("protected_endpoint",
+                                                          "GET") + self.get_correct_endpoints_for_method(
+            "public_endpoint", "GET")
+
+        for get_endpoint, _, _ in endpoints:
+            # Check if API Uses Session Management
+            for account in self.accounts:
+                if "api" in account and  account["api"] in get_endpoint:
+                    str_id = f"{account.get('id')}"
+                    get_endpoint = get_endpoint.replace("{id}", str_id)
+
+                    prompts = self.test_session_management(get_endpoint, account, prompts)
+
+
+                else:
+                    if "id}" in get_endpoint:
+                        get_endpoint = self.replace_placeholders_with_1(get_endpoint, account.get("id"))
+                    prompts = self.test_session_management(get_endpoint, account, prompts)
+
+                if self.login_endpoint:
+                    for login in self.login_endpoint:
+
+                        login_path, login_schema = self.get_path_and_schema(login)
+                        if login_schema is None:
+                            continue
+                        if "api" in account and account["api"] in login_path:
+                            str_id = f"{account.get('id')}"
+                            login_path = login_path.replace("{id}", str_id)
+
+                            prompts = self.test_session_hijacking(login_path, get_endpoint, login_schema, account,
+                                                                  prompts)
+
+                        else:
+                            prompts = self.test_session_hijacking(login_path, get_endpoint, login_schema, account,
+                                                                  prompts)
+                        prompts = self.test_sessions_vulnerabilitiy(login_path, login_schema, account, prompts)
+
+                        prompts = self.test_cookies(login_path, login_schema, prompts)
+
+        return prompts
+
+    def test_session_management(self, get_endpoint, account, prompts):
+        prompts = prompts + [
+            {
+                "objective": "Check if API Uses Session Management",
+                "steps": [
+                    f"Send a GET request to {get_endpoint} without authentication.",
+
+                ],
+                "path": [get_endpoint],
+                "token": [account.get("token")],
+                "expected_response_code": [
+                    "200 OK if no authentication is required, and the endpoint provides public data.",
+                    "401 Unauthorized if authentication is required and no session token or cookie is issued."
+                ],
+                "security": [
+                    "Observe the response headers to check if any session token or cookie is issued.",
+                    "Inspect both request and response headers for session-related identifiers."
+                    "Ensure that session tokens are not issued unnecessarily, as this could expose vulnerabilities or allow unintended session management."
+                ]
+            }]
+        return prompts
+
+    def test_session_hijacking(self, login_path, get_endpoint, login_schema, account, prompts):
+        prompts = prompts + [
+
+            # This prompt tests if the API uses session management by observing session tokens or cookies in responses after a GET request.
+            # Test for Session Hijacking Vulnerability
+            {
+                "objective": "Test for Session Hijacking Vulnerability",
+                "steps": [
+                    f"Send a POST request to {login_path} with valid credentials {self.get_credentials(login_schema, login_path)} to obtain a session token or cookie.",
+                    f"Send a GET request to {get_endpoint} with the session token or cookie and verify if access is granted without re-authentication."
+                ],
+                "path": [login_path, get_endpoint],
+                "expected_response_code": [
+                    "401 Unauthorized if the session token is invalid or the system detects unauthorized usage.",
+                    "403 Forbidden if the API identifies and blocks an unauthorized attempt to use a session token from a different environment or device."
+                ],
+                "token": ["", account.get("token"), account.get("token")],
+                "security": [
+                    "Ensure that the API has mechanisms such as IP binding or device fingerprinting to detect and prevent session hijacking attempts."
+                ]
+            }
+            # This prompt tests for session hijacking by attempting to reuse a session token from a different environment.
+            ,
+            # Check Session Timeout Settings
+            {
+                "objective": "Check Session Timeout Settings",
+                "steps": [
+                    f"Start a session by sending a POST request to {login_path} based on schema {login_schema} and obtaining a session identifier.",
+                    "Leave the session idle for the defined timeout period.",
+                    f"Send a GET request to {get_endpoint} using the same session token or cookie.",
+                ],
+                "token": ["", account.get("token"), account.get("token")],
+                "path": [login_path, get_endpoint, get_endpoint],
+                "expected_response_code": [
+                    "401 Unauthorized if the session has expired and the token is rejected due to inactivity.",
+                    "403 Forbidden if the API enforces access denial due to an expired session."
+                ],
+                "security": [
+                    "Verify if the session has expired and access is rejected."
+                    "Ensure that session timeout settings are enforced to minimize the risk of unauthorized access due to prolonged inactivity."
+                ]
+            }
+
+            # This prompt tests session timeout by verifying if a session expires after inactivity.
+            ,
+
+        ]
+
+        return prompts
+
+    def test_cookies(self, login_path, login_schema, prompts, account=None):
+        if account:
+
+            prompts.append(  # Check for SameSite Attribute on Cookies
+                {
+                    "objective": "Check for SameSite Attribute on Cookies",
+                    "steps": [
+                        f"Send a POST request to {login_path} based on schema {login_schema} with account {account} to authenticate and obtain a session cookie.",
+
+                    ],
+                    "path": [login_path],
+                    "token": [""],
+                    "expected_response_code": [
+                        "200 OK if the session cookie is successfully issued with appropriate attributes.",
+                    ],
+                    "security": [
+                        "Inspect the cookie attributes in the response headers to confirm that the SameSite attribute is set.",
+                        "Verify if the SameSite attribute is set to Strict or Lax, ensuring restricted cookie transmission across sites.",
+                        "Ensure that the SameSite attribute is configured to mitigate CSRF attacks by limiting cookie transmission.",
+                        "Verify that cookies with SameSite set to None also include the Secure attribute to prevent transmission over insecure connections."
+                    ]
+                }
+
+                # This prompt checks if the API enforces CSRF protection by setting the SameSite attribute on cookies.
+            )
+        else:
+            prompts = prompts + [  # Test for Secure Session Cookie Settings
+                {
+                    "objective": "Test for Secure Session Cookie Settings",
+                    "steps": [
+                        f"Send a POST request to {login_path} to obtain a session cookie based on schema {self.get_credentials(login_schema, login_path)} .",
+                    ],
+                    "token": [""],
+                    "path": [login_path],
+
+                    "expected_response_code": [
+                        "200 OK if the session cookie includes secure attributes such as HttpOnly and Secure."
+                    ],
+                    "security": [
+                        "Inspect the session token or cookie in the response headers to verify the inclusion of secure attributes such as HttpOnly and Secure."
+                        "Ensure that session cookies are configured with secure attributes to prevent unauthorized access or transmission of sensitive session information."
+                    ]
+                }]
+
+        # This prompt tests for secure session cookie attributes, such as HttpOnly and Secure, to protect against client-side access and enforce secure transmission.
+        return prompts
+
+    def test_sessions_vulnerabilitiy(self, login_path, login_schema, account, prompts):
+        endpoints = self.get_correct_endpoints_for_method("protected_endpoint",
+                                                          "POST") + self.get_correct_endpoints_for_method(
+            "public_endpoint", "POST")
+
+        for post_endpoint, _, _ in endpoints:
+            prompts = prompts + [
+
+                # Examine Session Handling Mechanism for Session Fixation Vulnerability
+                {
+                    "objective": "Examine Session Handling Mechanism for Session Fixation Vulnerability",
+                    "steps": [
+                        f"Authenticate and obtain a session identifier (token or cookie) by sending a POST request to {login_path} based on schema {self.get_credentials(login_schema, login_path)} .",
+                        f"Reuse the same session identifier across different sessions or user accounts by sending additional POST requests to {post_endpoint} with this session ID.",
+                    ],
+                    "token": ["", account.get("token")],
+                    "path": [login_path, post_endpoint],
+
+                    "expected_response_code": [
+                        "401 Unauthorized if the reused session identifier is invalid or has been appropriately invalidated.",
+                        "403 Forbidden if the API detects misuse or an unauthorized attempt to reuse the session."
+                    ],
+                    "security": [
+                        "Check if the session ID remains valid across multiple sessions or accounts."
+                        "Verify that the API generates new session identifiers upon login or account change to prevent session fixation attacks."
+                    ]
+                }
+
+                # This prompt tests for session fixation by reusing a session identifier across multiple login attempts.
+                ,
+
+            ]
+
+        return prompts
+
